### 2025-07-31 #8685 @kyleniemeyer
@JakobBD one minor suggestion: I did see your test suite and continuous integration set up through GitHub Actions. It might help reviewers and users if you add a badge to the top of the README that reports the status of the CI (e.g., "tests passing"); later, you can also add a badge with DOI for the JOSS citation, among others.
