### 2025-08-04 890abef
fix doi prefix issues
