+++ paper/paper.bib
@@ -7,7 +7,7 @@
     number = {3},
     pages = {446-458},
     keywords = {dynamic modeling, material flow analysis (MFA), open science, Python, software, substance flow analysis (SFA)},
-    doi = {https://doi.org/10.1111/jiec.12952},
+    doi = {10.1111/jiec.12952},
     url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/jiec.12952},
     eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/jiec.12952},
     year = {2020}
@@ -41,7 +41,7 @@ volume = {19},
 number = {5},
 pages = {728-741},
 keywords = {accounting, bipartite graphs, industrial ecology model, societal metabolism, supply and use tables, system of national accounts (SNA)},
-doi = {https://doi.org/10.1111/jiec.12306},
+doi = {10.1111/jiec.12306},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/jiec.12306},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/jiec.12306},
 year = {2015}
@@ -71,7 +71,7 @@ volume = {22},
 number = {2},
 pages = {434-442},
 keywords = {education and training, software, learning by doing, economy-wide material flow analysis (EW-MFA), urban metabolism, environmental education},
-doi = {https://doi.org/10.1111/jiec.12553},
+doi = {10.1111/jiec.12553},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/jiec.12553},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/jiec.12553},
 year = {2018}
