### 2025-08-04 #8685 @JakobBD
Thank you @kyleniemeyer!

I have added the badges and fixed DOI issues.

As reviewers, I suggest the following from the list of JOSS reviewers:
`konstantinstadler`
`Haitham-ghaida`
And from outside:
`cmutel`
`michaelweinold`
`TimoDiepers` ([submitted](https://github.com/openjournals/joss-reviews/issues/7981) a JOSS paper)

Looking forward to the review!
