### 2025-10-06 5930870
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/mwouts/jupytext: v1.17.2 → v1.17.3](https://github.com/mwouts/jupytext/compare/v1.17.2...v1.17.3)
- https://github.com/psf/black → https://github.com/psf/black-pre-commit-mirror
- [github.com/psf/black-pre-commit-mirror: 25.1.0 → 25.9.0](https://github.com/psf/black-pre-commit-mirror/compare/25.1.0...25.9.0)
- [github.com/pre-commit/pre-commit-hooks: v5.0.0 → v6.0.0](https://github.com/pre-commit/pre-commit-hooks/compare/v5.0.0...v6.0.0)
