### 2025-10-08 1f2c709
add __str__ methods for dimensions, dimension sets and flodym arrays
