### 2025-10-08 ae9e1a7
restructure stock computation especially for non-contiguous lifetimes
