+++ flodym/stocks.py
@@ -41,6 +41,7 @@ class Stock(PydanticBaseModel):
     """Process the stock is associated with, if any. Needed for example for the mass balance."""
     time_letter: str = "t"
     """Letter of the time dimension in the dimensions set, to make sure it's the first one."""
+    _t: UnevenTimeDim = None
 
     @model_validator(mode="after")
     def validate_stock_arrays(self):
@@ -70,6 +71,11 @@ class Stock(PydanticBaseModel):
             raise ValueError(
                 f"Time dimension must be the first dimension, i.e. time_letter (now {self.time_letter}) must be the first letter in dims.letters (now {self.dims.letters[0]})."
             )
+        return self
+
+    @model_validator(mode="after")
+    def init_t(self):
+        self._t = UnevenTimeDim(dim=self.dims[self.time_letter])
         return self
 
     @abstractmethod
@@ -115,6 +121,17 @@ class Stock(PydanticBaseModel):
         )  # stock_change(t) = stock(t) - stock(t-1)
         return self.inflow.values - self.outflow.values - dsdt
 
+    def _to_whole_period(self, annual_flow: np.ndarray) -> np.ndarray:
+        """multiply annual flow by interval length to get flow over whole period.
+        """
+        return np.einsum("t...,t->t...", annual_flow, self._t.interval_lengths)
+
+    def _to_annual(self, whole_period_flow: np.ndarray) -> np.ndarray:
+        """divide flow over whole period by interval length to get annual flow
+        """
+        return np.einsum("t...,t->t...", whole_period_flow, 1. / self._t.interval_lengths)
+
+
 
 class SimpleFlowDrivenStock(Stock):
     """Given inflows and outflows, the stock can be calculated without a lifetime model or cohorts."""
@@ -128,11 +145,9 @@ class SimpleFlowDrivenStock(Stock):
 
     def compute(self):
         self._check_needed_arrays()
-        annual_net_inflow = self.inflow - self.outflow
-        t = UnevenTimeDim(dim=self.dims[self.time_letter])
-        int_len = FlodymArray(dims=self.dims[self.time_letter,], values=t.interval_lengths)
-        net_inflow = annual_net_inflow * int_len
-        self.stock[...] = net_inflow.apply(np.cumsum, kwargs={"axis": 0})
+        annual_net_inflow = self.inflow.values - self.outflow.values
+        net_inflow_whole_period = self._to_whole_period(annual_net_inflow)
+        self.stock.values[...] = np.cumsum(net_inflow_whole_period, axis=0)
 
 
 class DynamicStockModel(Stock):
@@ -215,8 +230,10 @@ class InflowDrivenDSM(DynamicStockModel):
         self._compute_outflow()
 
     def _compute_stock(self):
+        # for non-contiguous years, yearly inflow is multiplied with time interval length
+        inflow_per_period = self._to_whole_period(self.inflow.values)
         self._stock_by_cohort = np.einsum(
-            "c...,tc...->tc...", self.inflow.values, self.lifetime_model.sf
+            "c...,tc...->tc...", inflow_per_period, self.lifetime_model.sf
         )
         self.stock.values[...] = self._stock_by_cohort.sum(axis=1)
 
@@ -259,36 +276,47 @@ class StockDrivenDSM(DynamicStockModel):
         where A is the survival function matrix, x is the inflow vector, and b is the stock vector.
         """
         if self.solver == "manual":
-            self._compute_cohorts_and_inflow_manual()
+            self._compute_inflow_manual()
         elif self.solver == "lapack":
-            self._compute_cohorts_and_inflow_lapack()
+            self._compute_inflow_lapack()
         else:
[…]
