### 2025-10-08 b8e03c8
add __str__ method for stocks
