+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "flodym"
-version = "0.5.0"
+version = "0.5.1"
 authors = [
     {name = "Jakob Dürrwächter"},
     {name = "Sally Dacie"},
