### 2025-10-08 b3c5439
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
