+++ pyproject.toml
@@ -6,6 +6,7 @@ authors = [
     {name = "Sally Dacie"},
     {name = "Merlin Hosak"},
     {name = "Bennet Weiß"},
+    {name = "Leonie Schweiger"},
 ]
 readme = "README.md"
 requires-python = ">= 3.10"
