### 2025-10-15 c44401b
Enable DimensionSet.index() to accept both letters and names
Co-authored-by: JakobBD <44020564+JakobBD@users.noreply.github.com>
