### 2025-10-15 e0f40cd
Move Dimension import to top of test file
Co-authored-by: JakobBD <44020564+JakobBD@users.noreply.github.com>
