### 2025-10-15 74f30d2
Apply black formatting to dimensions.py
Co-authored-by: JakobBD <44020564+JakobBD@users.noreply.github.com>
