### 2025-10-15 5d63a26
Enable ArrayPlotter to accept dimension letters and names
Co-authored-by: JakobBD <44020564+JakobBD@users.noreply.github.com>
