### 2025-10-17 a531aa8
extend array plotter tests
