### 2025-10-17 3f5c1ad
Fix ArrayPlotter._plot_subplot to handle dimension letters
Co-authored-by: JakobBD <44020564+JakobBD@users.noreply.github.com>
