### 2025-10-17 09f2fb6
Add tests and documentation for dimension letters as column names
Co-authored-by: JakobBD <44020564+JakobBD@users.noreply.github.com>
