### 2025-10-20 7c8e324
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
