### 2025-10-29 2edd83c
rename internal dimension porperty _dict to _full_maping
