+++ flodym/dimensions.py
@@ -148,7 +148,7 @@ class DimensionSet(PydanticBaseModel):
         return self
 
     @property
-    def _dict(self) -> Dict[str, Dimension]:
+    def _full_mapping(self) -> Dict[str, Dimension]:
         """Contains mappings.
 
         letter --> dim object and name --> dim object
@@ -170,7 +170,7 @@ class DimensionSet(PydanticBaseModel):
         if isinstance(key, tuple):
             return self.get_subset(key)
         if isinstance(key, str):
-            return self._dict[key]
+            return self._full_mapping[key]
         elif isinstance(key, int):
             return self.dim_list[key]
         else:
@@ -180,7 +180,7 @@ class DimensionSet(PydanticBaseModel):
         return iter(self.dim_list)
 
     def __contains__(self, key: str) -> bool:
-        return key in self._dict
+        return key in self._full_mapping
 
     def size(self, key: str):
         """get the number of items in a dimension
@@ -188,7 +188,7 @@ class DimensionSet(PydanticBaseModel):
         Args:
             key (str): the name or letter of the dimension to get the size of
         """
-        return self._dict[key].len
+        return self._full_mapping[key].len
 
     @property
     def shape(self) -> tuple[int]:
@@ -207,7 +207,7 @@ class DimensionSet(PydanticBaseModel):
         """
         subset = copy(self)
         if dims is not None:
-            subset.dim_list = [self._dict[dim_key] for dim_key in dims]
+            subset.dim_list = [self._full_mapping[dim_key] for dim_key in dims]
         return subset
 
     def expand_by(self, added_dims: list[Dimension]) -> "DimensionSet":
@@ -228,7 +228,7 @@ class DimensionSet(PydanticBaseModel):
         Returns:
             None if inplace=True, otherwise a new DimensionSet with the dimension removed
         """
-        dim_to_drop = self._dict[key]
+        dim_to_drop = self._full_mapping[key]
         if inplace:
             self.dim_list.remove(dim_to_drop)
             return
@@ -320,7 +320,7 @@ class DimensionSet(PydanticBaseModel):
         Args:
             key (str): The name or letter of the dimension to get the index of
         """
-        dim = self._dict[key]
+        dim = self._full_mapping[key]
         return self.dim_list.index(dim)
 
     def __str__(self):
