### 2025-11-11 816bff6
add DimensionSet operators
