### 2025-11-12 0e583d0
compute processes, whole implementation
