+++ flodym/flow_helper.py
@@ -26,11 +26,12 @@ def make_empty_flows(
     """
     flows = {}
     for flow_definition in flow_definitions:
-        try:
-            from_process = processes[flow_definition.from_process_name]
-            to_process = processes[flow_definition.to_process_name]
-        except KeyError:
-            raise KeyError(f"Missing process required by flow definition {flow_definition}.")
+        if flow_definition.from_process_name not in processes:
+            raise KeyError(f"Process name {flow_definition.from_process_name} given in flow definition is not found in processes.")
+        from_process = processes[flow_definition.from_process_name]
+        if flow_definition.to_process_name not in processes:
+            raise KeyError(f"Process name {flow_definition.to_process_name} given in flow definition is not found in processes.")
+        to_process = processes[flow_definition.to_process_name]
         if flow_definition.name_override is not None:
             name = flow_definition.name_override
         else:
