### 2025-11-12 c8a7975
add is_set property to flodym arrays.
