+++ flodym/mfa_system.py
@@ -302,6 +302,15 @@ class MFASystem(PydanticBaseModel):
         if all_good:
             logging.info(f"Success - No negative flows or NaN values in {self.__class__.__name__}")
 
+    def mark_all_unset(self):
+        """Mark all parameters as unset."""
+        for flow in self.flows.values():
+            flow.mark_unset()
+        for stock in self.stocks.values():
+            stock.inflow.mark_unset()
+            stock.outflow.mark_unset()
+            stock.stock.mark_unset()
+
     @staticmethod
     def _error_or_warning(message: str, raise_error: bool) -> bool:
         if raise_error:
