### 2025-11-12 f0889fc
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
