### 2025-11-12 17a2f3d
add flexible stock, switching between inflow driven and stock driven; delete lapack solver for stock driven
