### 2025-11-12 62d465f
make process definition object; store links to inflows and outflows in processes
