+++ flodym/mfa_definition.py
@@ -126,6 +126,13 @@ class ParameterDefinition(DefinitionWithDimLetters):
     """Name of the parameter."""
 
 
+class ProcessDefinition(PydanticBaseModel):
+    id: int
+    """ID of the process."""
+    name: str
+    """Name of the process."""
+
+
 class MFADefinition(PydanticBaseModel):
     """All the information needed to define an MFA system, compiled of lists of definition objects."""
 
@@ -133,7 +140,7 @@ class MFADefinition(PydanticBaseModel):
 
     dimensions: List[DimensionDefinition]
     """List of definitions of dimensions used in the model."""
-    processes: List[str]
+    processes: List[str|ProcessDefinition]
     """List of process names used in the model."""
     flows: List[FlowDefinition]
     """List of definitions of flows used in the model."""
