### 2025-11-12 b3f7995
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
