### 2025-11-13 701082f
add config file; refactor mass balance checks
