+++ flodym/mfa_system.py
@@ -24,6 +24,7 @@ from .data_reader import (
     ExcelDimensionReader,
     ExcelParameterReader,
 )
+from .config import config, ErrorBehavior, handle_error
 
 
 class MFASystem(PydanticBaseModel):
@@ -180,86 +181,50 @@ class MFASystem(PydanticBaseModel):
         dims = self.dims.get_subset(dim_letters)
         return FlodymArray(dims=dims, **kwargs)
 
-    def _get_mass_balance(self) -> Dict[str, FlodymArray]:
-        """Calculate the mass balance for each process, by summing the contributions.
-        - all flows entering (positive)
-        - all flows leaving (negative)
-        - the stock change of the process (negative)
-
-        Returns:
-            A dictionary mapping process names to their mass balance contributions.
-            Each contribution is a :py:class:`flodym.FlodymArray` with dimensions common to all contributions.
-        """
-        contributions = {p: [] for p in self.processes.keys()}
-
-        # Add flows to mass balance
-        for flow in self.flows.values():
-            contributions[flow.from_process.name].append(-flow)  # Subtract flow from start process
-            contributions[flow.to_process.name].append(flow)  # Add flow to end process
-
-        # Add stock changes to the mass balance
-        for stock in self.stocks.values():
-            if stock.process is None:  # not connected to a process
-                continue
-            stock_change = stock.inflow - stock.outflow
-            #    sum(flows_to) - sum(flows_from) = stock_change
-            # => sum(flows_to) - sum(flows_from) - stock_change = 0
-            # => stock_change is subtracted from the process
-            contributions[stock.process.name].append(-stock_change)
-            # system_mass_change = sum(stock_changes),
-            # where the sysenv process mass balance is the negative system_mass_change:
-            # system_mass_change = flows_into_system - flows_out_of_system
-            #                    = sum(flows_from_sysenv) - sum(flows_to_sysenv)
-            # So stock change is accounted to sysenv process with opposite sign as to other
-            # processes => added instead of subtracted.
-            contributions["sysenv"].append(stock_change)
-
-        return {p_name: sum(parts) for p_name, parts in contributions.items()}
-
-    @property
-    def _absolute_float_precision(self) -> float:
-        """The numpy float precision, multiplied by the maximum absolute flow or stock value."""
-        max_flow_value = max([np.max(np.abs(f.values)) for f in self.flows.values()])
-        max_stock_value = max([np.max(np.abs(s.stock.values)) for s in self.stocks.values()])
-        epsilon = np.finfo(next(iter(self.flows.values())).values.dtype).eps
-        return epsilon * max(max_flow_value, max_stock_value)
-
-    def check_mass_balance(self, tolerance=None, raise_error: bool = True):
+    def check_mass_balance(self, error_behavior: ErrorBehavior = None):
         """Compute mass balance, and check whether it is within a certain tolerance.
         Throw an error if it isn't.
 
         Args:
-            tolerance (float, optional): The tolerance for the mass balance check.
-                If None, defaults to 100 times the numpy float precision,
-                multiplied by the maximum absolute flow or stock value.
-            raise_error (bool): If True, raises an error if the mass balance check fails.
-                Else, logs a warning and continues execution.
+            error_behavior (ErrorBehavior): What to do if the mass balance check fails.
+                If None, takes the global config setting, which defaults to raising an error.
         """
 
         logging.info(f"Checking mass balance of {self.__class__.__name__} object...")
 
-        if tolerance is None:
-            tolerance = 100 * self._absolute_float_precision
+        self.processes["sysenv"].check_mass_balance(tolerance=self.tolerance, error_behavior=error_behavior, mass_change_target=-self.total_stock_change,)
[…]
