+++ flodym/processes.py
@@ -11,7 +11,6 @@ from .dimensions import DimensionSet
 from .config import config, handle_error, ErrorBehavior
 
 
-
 class Process(PydanticBaseModel, arbitrary_types_allowed=True):
     """Processes serve as nodes for the MFA system layout definition.
     Flows are defined between two processes. Stocks are connected to a process.
@@ -96,29 +95,29 @@ class Process(PydanticBaseModel, arbitrary_types_allowed=True):
             raise ValueError("Direction must be 'in' or 'out'.")
 
     def known_flows(self, side: str) -> Dict[str, Flow]:
-        return {
-            name: flow for name, flow in self.flows(side).items() if flow.is_set
-        }
+        return {name: flow for name, flow in self.flows(side).items() if flow.is_set}
 
     def unknown_flows(self, side: str) -> Dict[str, Flow]:
-        return {
-            name: flow for name, flow in self.flows(side).items() if not flow.is_set
-        }
+        return {name: flow for name, flow in self.flows(side).items() if not flow.is_set}
 
     def get_sum_inflows(self) -> FlodymArray:
         """Sum of all inflows to the process."""
         if not self.is_computed:
-            raise ValueError(f"Cannot compute sum_inflows for process {self.name} before it is computed.")
+            raise ValueError(
+                f"Cannot compute sum_inflows for process {self.name} before it is computed."
+            )
         if not self.inflows:
-            return 0.
+            return 0.0
         return sum(self.inflows.values())
 
     def get_sum_outflows(self) -> FlodymArray:
         """Sum of all outflows from the process."""
         if not self.is_computed:
-            raise ValueError(f"Cannot compute sum_outflows for process {self.name} before it is computed.")
+            raise ValueError(
+                f"Cannot compute sum_outflows for process {self.name} before it is computed."
+            )
         if not self.outflows:
-            return 0.
+            return 0.0
         return sum(self.outflows.values())
 
     def get_net_inflow(self):
@@ -130,11 +129,11 @@ class Process(PydanticBaseModel, arbitrary_types_allowed=True):
         if self.inflows:
             max_inflow = max([f._absolute_float_precision for f in self.inflows.values()])
         else:
-            max_inflow = 0.
+            max_inflow = 0.0
         if self.outflows:
             max_outflow = max([f._absolute_float_precision for f in self.outflows.values()])
         else:
-            max_outflow = 0.
+            max_outflow = 0.0
         return max(max_inflow, max_outflow)
 
     @property
@@ -144,7 +143,12 @@ class Process(PydanticBaseModel, arbitrary_types_allowed=True):
             return config.absolute_tolerance
         return config.relative_tolerance * self._absolute_float_precision
 
-    def check_mass_balance(self, tolerance: float = None, error_behavior: ErrorBehavior = None, mass_change_target: FlodymArray = None):
+    def check_mass_balance(
+        self,
+        tolerance: float = None,
+        error_behavior: ErrorBehavior = None,
+        mass_change_target: FlodymArray = None,
+    ):
         """Compute mass balance, and check whether it is within a certain tolerance.
         Throw an error if it isn't.
 
@@ -156,16 +160,18 @@ class Process(PydanticBaseModel, arbitrary_types_allowed=True):
         """
 
         if self.id == 0 and mass_change_target is None:
-            logging.info(f"mass_change_target not given. Skip mass balance check for system environment.")
+            logging.info(
+                f"mass_change_target not given. Skip mass balance check for system environment."
+            )
             return
 
         logging.info(f"Checking mass balance of {self.__class__.__name__} object...")
 
         if mass_change_target is None:
-            mass_change_target = 0. if self.stock is None else self.stock.net_inflow
[…]
