### 2025-11-13 89830b9
add functionality to auto-compute processes from definition structure outline
