+++ flodym/__init__.py
@@ -14,7 +14,12 @@ from flodym.flodym_arrays import (
     StockArray as StockArray,
     Flow as Flow,
 )
-from flodym.processes import Process as Process, make_processes as make_processes
+from flodym.processes import (
+    Process as Process,
+    make_processes as make_processes,
+    set_process_parameters as set_process_parameters,
+    UnderdeterminedError as UnderdeterminedError,
+)
 from flodym.stocks import (
     Stock as Stock,
     SimpleFlowDrivenStock as SimpleFlowDrivenStock,
