+++ flodym/mfa_definition.py
@@ -131,6 +131,12 @@ class ProcessDefinition(PydanticBaseModel):
     """ID of the process."""
     name: str
     """Name of the process."""
+    inflow_shares: Optional[Dict[str, str]] = None
+    """parameter names to use as inflow shares, keyed by from_process name of inflow"""
+    outflow_shares: Optional[Dict[str, str]] = None
+    """parameter names to use as outflow shares, keyed by to_process name of outflow"""
+    dimension_splitter: Optional[str] = None
+    """Name of the parameter to use for splitting the total process flow into new dimensions."""
 
 
 class MFADefinition(PydanticBaseModel):
