+++ flodym/processes.py
@@ -11,6 +11,14 @@ from .dimensions import DimensionSet
 from .config import config, handle_error, ErrorBehavior
 
 
+class UnderdeterminedError(Exception):
+    """Exception raised when a process is underdetermined."""
+
+    def __init__(self, process: "Process", message: str):
+        message = f"Cannot compute process '{process.name}' with ID {process.id}, as it is underdetermined. \n" + message
+        super().__init__(message)
+        self.message = message
+
 class Process(PydanticBaseModel, arbitrary_types_allowed=True):
     """Processes serve as nodes for the MFA system layout definition.
     Flows are defined between two processes. Stocks are connected to a process.
@@ -29,7 +37,12 @@ class Process(PydanticBaseModel, arbitrary_types_allowed=True):
     """Inflows to the process, keyed by linked process name."""
     _outflows: Dict[str, Flow] = {}
     """Outflows from the process, keyed by linked process name."""
+    dimension_splitter: Optional["FlodymArray"] = None
     stock: Optional[Stock] = None
+    _inflow_shares: Dict[str, "FlodymArray"] = {}
+    _outflow_shares: Dict[str, "FlodymArray"] = {}
+    _total: FlodymArray = None
+    _was_overdetermined: bool = None
 
     @model_validator(mode="after")
     def check_id0(self):
@@ -77,6 +90,30 @@ class Process(PydanticBaseModel, arbitrary_types_allowed=True):
         else:
             raise ValueError(f"In process {self.name}: No outflow to process '{to_process}' found.")
 
+    def add_inflow_share(self, from_process: str, share: FlodymArray):
+        if from_process not in self._inflows:
+            raise ValueError(f"In process {self.name}: Cannot add inflow share from process {from_process}, as no such inflow found.")
+        self._inflow_shares[from_process] = share
+
+    def add_outflow_share(self, to_process: str, share: FlodymArray):
+        if to_process not in self._outflows:
+            raise ValueError(f"In process {self.name}: Cannot add outflow share to process {to_process}, as no such outflow found.")
+        self._outflow_shares[to_process] = share
+
+    def remove_inflow_share(self, from_process: str):
+        """Remove the inflow share for a given process."""
+        if from_process in self._inflow_shares:
+            del self._inflow_shares[from_process]
+        else:
+            raise ValueError(f"In process {self.name}: No inflow share for process '{from_process}' found.")
+
+    def remove_outflow_share(self, to_process: str):
+        """Remove the outflow share for a given process."""
+        if to_process in self._outflow_shares:
+            del self._outflow_shares[to_process]
+        else:
+            raise ValueError(f"In process {self.name}: No outflow share for process '{to_process}' found.")
+
     @property
     def is_computed(self) -> bool:
         """Check if the process has been computed."""
@@ -86,6 +123,91 @@ class Process(PydanticBaseModel, arbitrary_types_allowed=True):
             return False
         return True
 
+    def compute(self, on_underdetermined: ErrorBehavior = "error", recursive: bool=False) -> None:
+        """Compute all unknown flows of the process, based on topological information and known in/outflows.
+        This covers most cases of MFA processes, but not all.
+        Sometimes manual computations are still necessary.
+        For example, if there is a 2x2 equation system to solve (see example 1).
+
+        Topological information needed for this includes:
+        - the inflows and outflows
+        - the shares of some in- or outflows relative to the process total
+        - an array to multiply with to apply a dimension split
+
+        The computation assumes the following process structure:
+        - one or several inflows
+        - combined into a total
+        - possible dimension change
+        - one or several outflows
+        Though quite general, this applies some restrictions on the possible process layouts.
+        For example, no dimension expansion of individual in- or outflows through multiplication with shares is allowed.
+        However, this can be realized by adding an own process which only applies this dimension splitting.
+
[…]
