+++ flodym/processes.py
@@ -15,9 +15,13 @@ class UnderdeterminedError(Exception):
     """Exception raised when a process is underdetermined."""
 
     def __init__(self, process: "Process", message: str):
-        message = f"Cannot compute process '{process.name}' with ID {process.id}, as it is underdetermined. \n" + message
+        message = (
+            f"Cannot compute process '{process.name}' with ID {process.id}, as it is underdetermined. \n"
+            + message
+        )
         super().__init__(message)
         self.message = message
+
 
 class Process(PydanticBaseModel, arbitrary_types_allowed=True):
     """Processes serve as nodes for the MFA system layout definition.
@@ -92,12 +96,16 @@ class Process(PydanticBaseModel, arbitrary_types_allowed=True):
 
     def add_inflow_share(self, from_process: str, share: FlodymArray):
         if from_process not in self._inflows:
-            raise ValueError(f"In process {self.name}: Cannot add inflow share from process {from_process}, as no such inflow found.")
+            raise ValueError(
+                f"In process {self.name}: Cannot add inflow share from process {from_process}, as no such inflow found."
+            )
         self._inflow_shares[from_process] = share
 
     def add_outflow_share(self, to_process: str, share: FlodymArray):
         if to_process not in self._outflows:
-            raise ValueError(f"In process {self.name}: Cannot add outflow share to process {to_process}, as no such outflow found.")
+            raise ValueError(
+                f"In process {self.name}: Cannot add outflow share to process {to_process}, as no such outflow found."
+            )
         self._outflow_shares[to_process] = share
 
     def remove_inflow_share(self, from_process: str):
@@ -105,14 +113,18 @@ class Process(PydanticBaseModel, arbitrary_types_allowed=True):
         if from_process in self._inflow_shares:
             del self._inflow_shares[from_process]
         else:
-            raise ValueError(f"In process {self.name}: No inflow share for process '{from_process}' found.")
+            raise ValueError(
+                f"In process {self.name}: No inflow share for process '{from_process}' found."
+            )
 
     def remove_outflow_share(self, to_process: str):
         """Remove the outflow share for a given process."""
         if to_process in self._outflow_shares:
             del self._outflow_shares[to_process]
         else:
-            raise ValueError(f"In process {self.name}: No outflow share for process '{to_process}' found.")
+            raise ValueError(
+                f"In process {self.name}: No outflow share for process '{to_process}' found."
+            )
 
     @property
     def is_computed(self) -> bool:
@@ -123,7 +135,7 @@ class Process(PydanticBaseModel, arbitrary_types_allowed=True):
             return False
         return True
 
-    def compute(self, on_underdetermined: ErrorBehavior = "error", recursive: bool=False) -> None:
+    def compute(self, on_underdetermined: ErrorBehavior = "error", recursive: bool = False) -> None:
         """Compute all unknown flows of the process, based on topological information and known in/outflows.
         This covers most cases of MFA processes, but not all.
         Sometimes manual computations are still necessary.
@@ -164,7 +176,9 @@ class Process(PydanticBaseModel, arbitrary_types_allowed=True):
             logging.debug(f"Process {self.name} with ID {self.id} is already computed.")
             return
         if self.id == 0:
-            logging.debug(f"Process {self.name} with ID {self.id} is the system boundary and cannot be computed.")
+            logging.debug(
+                f"Process {self.name} with ID {self.id} is the system boundary and cannot be computed."
+            )
             return
         try:
 
@@ -226,8 +240,9 @@ class Process(PydanticBaseModel, arbitrary_types_allowed=True):
 
     def neither_known(self, side: str) -> List[str]:
         return [
[…]
