### 2025-11-13 ad741ab
add functionality for handling Dimensions and DimensionSets
