### 2025-11-14 a2e2903
rename and use stock mark_set / mark_unset methods
