+++ flodym/mfa_system.py
@@ -303,13 +303,11 @@ class MFASystem(PydanticBaseModel):
             logging.info(f"Success - No negative flows or NaN values in {self.__class__.__name__}")
 
     def mark_all_unset(self):
-        """Mark all parameters as unset."""
+        """Mark all stocks and flows as unset."""
         for flow in self.flows.values():
             flow.mark_unset()
         for stock in self.stocks.values():
-            stock.inflow.mark_unset()
-            stock.outflow.mark_unset()
-            stock.stock.mark_unset()
+            stock.mark_all_unset()
 
     @staticmethod
     def _error_or_warning(message: str, raise_error: bool) -> bool:
