### 2025-11-22 9ed0ef0
Label Sankey split flows with dim items
