+++ flodym/export/sankey.py
@@ -151,10 +151,11 @@ class PlotlySankeyPlotter(CustomNameDisplayer, PydanticBaseModel):
         if isinstance(color, tuple):
             split_flows_by = color[0]
             colors = color[1]
+            labels = self.mfa.dims[split_flows_by].items
             splitting_letter_tuple = (self.mfa.dims[split_flows_by].letter,)
             values = f_slice.sum_values_to(splitting_letter_tuple)
-            for v, c in zip(values, colors):
-                links.append(label=label, source=source, target=target, color=c, value=v)
+            for l, c, v in zip(labels, colors, values):
+                links.append(label=l, source=source, target=target, color=c, value=v)
         else:
             links.append(
                 label=label,
