### 2025-11-22 d63454a
Refactor PlotlySankeyPlotter._append_flow() for readability
