### 2025-11-27 232877c
some new operators, some more documentation in the howto
