+++ flodym/dimensions.py
@@ -3,6 +3,7 @@ from pydantic import BaseModel as PydanticBaseModel, Field, AliasChoices, model_
 from typing import Dict, Iterator, Optional
 import numpy as np
 import pandas as pd
+from deprecated import deprecated
 
 from .mfa_definition import DimensionDefinition
 
@@ -97,9 +98,18 @@ class Dimension(PydanticBaseModel):
         """Check if the items of this dimension are a superset of the items of another dimension."""
         return set(self.items).issuperset(other.items)
 
-    def to_set(self) -> "DimensionSet":
+    def as_dimset(self) -> "DimensionSet":
         """Convert the Dimension to a DimensionSet containing only this Dimension."""
         return DimensionSet(dim_list=[self])
+
+    def __add__(self, other: "Dimension") -> "DimensionSet":
+        """Addition operator to create a DimensionSet from two Dimensions."""
+        if isinstance(other, Dimension):
+            return DimensionSet(dim_list=[self, other])
+        elif isinstance(other, DimensionSet):
+            return self.as_dimset() + other
+        else:
+            raise TypeError("Can only add Dimension to Dimension or DimensionSet.")
 
     def __str__(self):
         base = f"Dimension '{self.name}' ('{self.letter}'); "
@@ -183,7 +193,9 @@ class DimensionSet(PydanticBaseModel):
     def __iter__(self) -> Iterator[Dimension]:
         return iter(self.dim_list)
 
-    def __contains__(self, key: str) -> bool:
+    def __contains__(self, key: str | Dimension) -> bool:
+        if isinstance(key, Dimension):
+            key = key.letter
         return key in self._full_mapping
 
     def size(self, key: str):
@@ -209,6 +221,7 @@ class DimensionSet(PydanticBaseModel):
             subset.dim_list = [self._full_mapping[dim_key] for dim_key in dims]
         return subset
 
+    @deprecated(reason="Use addition operator '+' or 'append' method instead", version="0.5.2")
     def expand_by(self, added_dims: list[Dimension]) -> "DimensionSet":
         """Expands the DimensionSet by adding new dimensions to it."""
         if not all([dim.letter not in self.letters for dim in added_dims]):
@@ -216,6 +229,68 @@ class DimensionSet(PydanticBaseModel):
                 "DimensionSet already contains one or more of the dimensions to be added."
             )
         return DimensionSet(dim_list=self.dim_list + added_dims)
+
+    def _check_additional_dim(self, new_dim: Dimension):
+        """Checks the new_dim is a Dimension and that it is not already in the DimensionSet."""
+        if not isinstance(new_dim, Dimension):
+            raise TypeError("new_dim must be a Dimension object.")
+        if new_dim in self:
+            raise ValueError(
+                "New dimension can't have same letter as any of those already in DimensionSet, "
+                "as that would create ambiguity."
+            )
+
+    def append(self, new_dim: Dimension, inplace: bool = False) -> Optional["DimensionSet"]:
+        """Add a new dimension to the set at the end.
+
+        Args:
+            new_dim (Dimension): The new dimension to add
+            inplace (bool, optional): If True, the operation is performed in place, otherwise a new DimensionSet is returned. Defaults to False.
+
+        Returns:
+            None if inplace=True, otherwise a new DimensionSet with the new dimension added
+        """
+        self._check_additional_dim(new_dim)
+        if inplace:
+            self.dim_list.append(new_dim)
+            return
+        else:
+            return self + new_dim
+
+    def prepend(self, new_dim: Dimension, inplace: bool = False) -> Optional["DimensionSet"]:
+        """Add a new dimension to the set at the beginning.
+        Args:
[…]
