### 2025-11-27 8be11ca
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
