### 2025-11-28 4a219b3
bugfix in FlodymArray setitem for non-ellipsis index and non-FlodymArray value
