### 2025-11-28 bcd9f78
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
