+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "flodym"
-version = "0.5.1"
+version = "0.5.2"
 authors = [
     {name = "Jakob Dürrwächter"},
     {name = "Sally Dacie"},
