### 2025-12-08 bae5762
Add comprehensive tests for Dimension and DimensionSet functions
- Test Dimension.as_dimset() method
- Test Dimension.__add__() operator for Dimension+Dimension and Dimension+DimensionSet
- Test sum() operator with Dimensions (uses __add__)
- Test DimensionSet.append() with inplace and non-inplace modes
- Test DimensionSet.prepend() with inplace and non-inplace modes
- Test DimensionSet.remove() (alias for drop()) with both modes
- Test DimensionSet.insert() at various positions with both modes
- Test DimensionSet.__add__() operator for DimensionSet+DimensionSet and DimensionSet+Dimension
- Fi
[…]
