### 2025-12-08 603293b
Fix comment in test_dimension_sum_operator to clarify sum() behavior
Co-authored-by: JakobBD <44020564+JakobBD@users.noreply.github.com>
