### 2025-12-08 65e0bbf
remove sum() test, as it does not work as intended
