+++ tests/test_dimensions.py
@@ -330,18 +330,18 @@ def test_ndim_property():
 def test_dimension_as_dimset():
     """Test that as_dimset() converts a Dimension to a DimensionSet."""
     dim = Dimension(name="time", letter="t", items=[1990, 2000, 2010])
-    
+
     # Convert to DimensionSet
     dimset = dim.as_dimset()
-    
+
     # Verify it's a DimensionSet
     assert isinstance(dimset, DimensionSet)
-    
+
     # Verify it contains only the one dimension
     assert len(dimset) == 1
     assert dimset.letters == ("t",)
     assert dimset.names == ("time",)
-    
+
     # Verify the dimension is the same
     assert dimset[0] == dim
 
@@ -350,54 +350,27 @@ def test_dimension_add():
     """Test the __add__ operator for Dimension objects."""
     dim1 = Dimension(name="time", letter="t", items=[1990, 2000, 2010])
     dim2 = Dimension(name="place", letter="p", items=["World", "Moon"])
-    
+
     # Test Dimension + Dimension
     dimset = dim1 + dim2
     assert isinstance(dimset, DimensionSet)
     assert len(dimset) == 2
     assert dimset.letters == ("t", "p")
     assert dimset.names == ("time", "place")
-    
+
     # Test Dimension + DimensionSet
     dim3 = Dimension(name="material", letter="m", items=["steel", "aluminum"])
     dimset2 = dim1 + dim2
     dimset3 = dim3 + dimset2
-    
+
     assert isinstance(dimset3, DimensionSet)
     assert len(dimset3) == 3
     assert dimset3.letters == ("m", "t", "p")
     assert dimset3.names == ("material", "time", "place")
-    
+
     # Test adding invalid type
     with pytest.raises(TypeError):
         dim1 + "invalid"
-
-
-def test_dimension_sum_operator():
-    """Test that sum() works with Dimension objects using __add__."""
-    dim1 = Dimension(name="time", letter="t", items=[1990, 2000, 2010])
-    dim2 = Dimension(name="place", letter="p", items=["World", "Moon"])
-    dim3 = Dimension(name="material", letter="m", items=["steel", "aluminum"])
-    
-    # Test sum with list of Dimensions
-    # sum() defaults to start=0, but Dimension objects cannot be added to integers,
-    # so we must provide an empty DimensionSet as the start value
-    dimset = sum([dim1, dim2, dim3], start=DimensionSet(dim_list=[]))
-    
-    assert isinstance(dimset, DimensionSet)
-    assert len(dimset) == 3
-    assert dimset.letters == ("t", "p", "m")
-    assert dimset.names == ("time", "place", "material")
-    
-    # Test sum with single dimension
-    dimset_single = sum([dim1], start=DimensionSet(dim_list=[]))
-    assert isinstance(dimset_single, DimensionSet)
-    assert len(dimset_single) == 1
-    
-    # Test sum with empty list
-    dimset_empty = sum([], start=DimensionSet(dim_list=[]))
-    assert isinstance(dimset_empty, DimensionSet)
-    assert len(dimset_empty) == 0
 
 
 def test_dimensionset_append():
@@ -408,29 +381,29 @@ def test_dimensionset_append():
     ]
     dimset = DimensionSet(dim_list=dims)
     new_dim = Dimension(name="material", letter="m", items=["steel", "aluminum"])
-    
+
     # Test append with inplace=False (returns new DimensionSet)
     new_dimset = dimset.append(new_dim, inplace=False)
     assert new_dimset is not None
     assert len(new_dimset) == 3
     assert new_dimset.letters == ("t", "p", "m")
-    
+
     # Original should be unchanged
     assert len(dimset) == 2
[…]
