+++ .pre-commit-config.yaml
@@ -7,7 +7,7 @@ repos:
       additional_dependencies:
         - black==24.10.0 # should match hook
   - repo: https://github.com/psf/black-pre-commit-mirror
-    rev: 25.11.0
+    rev: 25.12.0
     hooks:
       - id: black
         language_version: python3.12
