### 2025-12-11 f18d93a
initial stocks first implementation
