+++ flodym/stocks.py
@@ -4,13 +4,13 @@ including flow-driven stocks and dynamic (lifetime-based) stocks.
 
 from abc import abstractmethod
 import numpy as np
-from pydantic import BaseModel as PydanticBaseModel, ConfigDict, model_validator
+from pydantic import BaseModel as PydanticBaseModel, ConfigDict, model_validator, Field
 from typing import Optional, Union
 import logging
 
 from .processes import Process
 from .flodym_arrays import StockArray, FlodymArray
-from .dimensions import DimensionSet
+from .dimensions import Dimension, DimensionSet
 from .lifetime_models import LifetimeModel, UnevenTimeDim
 
 
@@ -201,8 +201,31 @@ class DynamicStockModel(Stock):
     Can be input either as a LifetimeModel subclass, or as an instance of a
     LifetimeModel subclass. For available subclasses, see `flodym.lifetime_models`.
     """
+    cohort_dim: Optional[Dimension] = None
+
     _outflow_by_cohort: np.ndarray = None
     _stock_by_cohort: np.ndarray = None
+    _dims_cohort: DimensionSet = None
+    _initial_stock_dsm: "InflowByCohortDrivenDSM" = None
+    _initial_stock_year: int = None
+
+    @model_validator(mode="after")
+    def validate_cohort_dim(self):
+        if self.cohort_dim is not None:
+            t = self.dims[self.time_letter]
+            c = self.cohort_dim
+            if c.letter == t.letter or c.name == t.name:
+                raise ValueError(
+                    "Cohort dimension letter and name must be different from time dimension letter and name."
+                )
+            if c.items != t.items:
+                raise ValueError(
+                    "Cohort dimension size must be the same as time dimension size."
+                )
+            if c.letter in self.dims.letters or c.name in self.dims.names:
+                raise ValueError("Cohort dimension must not be part of the stock dimensions.")
+            self._dims_cohort = t + c + self.dims.drop(self.time_letter, inplace=False)
+        return self
 
     @model_validator(mode="after")
     def init_cohort_arrays(self):
@@ -223,6 +246,10 @@ class DynamicStockModel(Stock):
     def _check_needed_arrays(self):
         self.lifetime_model._check_prms_set()
 
+    def _check_cohort_dim(self, application: str):
+        if self.cohort_dim is None:
+            raise ValueError(f"Cohort dimension must be provided at DSM initialization for {application} to work.")
+
     @property
     def _n_t(self) -> int:
         return list(self.shape)[0]
@@ -239,19 +266,85 @@ class DynamicStockModel(Stock):
     def _t_diag_indices(self) -> tuple:
         return np.diag_indices(self._n_t) + (slice(None),) * len(self._shape_no_t)
 
-    def get_outflow_by_cohort(self) -> np.ndarray:
+    def get_outflow_by_cohort(self) -> StockArray:
         """Outflow by cohort, i.e. the outflow of each production year at each time step."""
-        return self._outflow_by_cohort
+        return self._get_by_cohort_array(self._outflow_by_cohort, "outflow_by_cohort")
 
-    def get_stock_by_cohort(self) -> np.ndarray:
+    def get_stock_by_cohort(self) -> StockArray:
         """Stock by cohort, i.e. the stock of each production year at each time step."""
-        return self._stock_by_cohort
+        return self._get_by_cohort_array(self._stock_by_cohort, "stock_by_cohort")
+
+    def _get_by_cohort_array(self, values: np.ndarray, name: str) -> StockArray:
+        if self.cohort_dim is None:  # leave in for backwards compatibility
+            logging.warning(
+                f"Cohort dimension is not defined; {name} cannot be retrieved as FlodymArray. Returning raw ndarray instead."
+            )
[…]
