### 2025-12-11 2498d30
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
