### 2025-12-11 987706d
Fix bugs in DSM _check_needed_arrays and InflowByCohortDrivenDSM
Co-authored-by: JakobBD <44020564+JakobBD@users.noreply.github.com>
