### 2025-12-11 12cde4b
Add comprehensive tests for DSM functionality and fix bugs
Co-authored-by: JakobBD <44020564+JakobBD@users.noreply.github.com>
