+++ tests/test_stocks.py
@@ -4,7 +4,7 @@ import time
 
 from flodym.dimensions import Dimension, DimensionSet
 from flodym.flodym_arrays import StockArray
-from flodym.stocks import InflowDrivenDSM, StockDrivenDSM
+from flodym.stocks import InflowDrivenDSM, StockDrivenDSM, InflowByCohortDrivenDSM
 from flodym.lifetime_models import LogNormalLifetime
 
 dim_list = [
@@ -166,3 +166,290 @@ def test_unequal_time_steps():
     stocks_select = stocks[1]
     values_select = stocks_select.values[-10:]
     assert np.max(np.abs(values_all - values_select)) < 0.01
+
+
+def test_inflow_by_cohort_driven_dsm():
+    """Test InflowByCohortDrivenDSM basic functionality."""
+    # Create simple dimensions for testing
+    time_items = list(range(2000, 2011))
+    dims_test = DimensionSet(
+        dim_list=[
+            Dimension(name="time", letter="t", items=time_items, dtype=int),
+            Dimension(name="product", letter="p", items=["A", "B"], dtype=str),
+        ]
+    )
+    
+    # Create cohort dimension
+    cohort_dim = Dimension(name="cohort", letter="c", items=time_items, dtype=int)
+    
+    # Create lifetime model
+    lifetime_model = LogNormalLifetime(dims=dims_test, time_letter="t", mean=5, std=2)
+    
+    # Create InflowByCohortDrivenDSM
+    dsm = InflowByCohortDrivenDSM(
+        dims=dims_test,
+        cohort_dim=cohort_dim,
+        lifetime_model=lifetime_model,
+        time_letter="t",
+        name="test_dsm",
+    )
+    
+    # Set inflow by cohort - diagonal pattern (inflow only occurs in cohort year)
+    dsm.inflow_by_cohort.values[...] = 0.0
+    for i in range(len(time_items)):
+        dsm.inflow_by_cohort.values[i, i, :] = 1.0  # 1 unit per product per year
+    
+    # Compute the DSM
+    dsm.compute()
+    
+    # Check that inflow was computed correctly (diagonal extraction)
+    assert dsm.inflow.is_set
+    assert dsm.stock.is_set
+    assert dsm.outflow.is_set
+    
+    # Verify inflow values are approximately 1.0 at each time step
+    assert np.allclose(dsm.inflow.values, 1.0)
+    
+    # Verify stock balance
+    dsm.check_stock_balance()
+
+
+def test_inflow_driven_dsm_with_initial_stock():
+    """Test InflowDrivenDSM with initial stock set via set_initial_stock."""
+    # Create dimensions
+    time_items = list(range(2000, 2021))
+    dims_test = DimensionSet(
+        dim_list=[
+            Dimension(name="time", letter="t", items=time_items, dtype=int),
+            Dimension(name="product", letter="p", items=["A", "B"], dtype=str),
+        ]
+    )
+    
+    # Create cohort dimension (required for initial stock functionality)
+    cohort_dim = Dimension(name="cohort", letter="c", items=time_items, dtype=int)
+    
+    # Create lifetime model
+    lifetime_model = LogNormalLifetime(dims=dims_test, time_letter="t", mean=8, std=3)
+    
+    # Create InflowDrivenDSM with cohort_dim
+    dsm = InflowDrivenDSM(
+        dims=dims_test,
+        cohort_dim=cohort_dim,
+        lifetime_model=lifetime_model,
+        time_letter="t",
+        name="test_dsm_initial",
+    )
+    
+    # Set initial stock at year 2005
+    initial_year = 2005
+    # Initial stock dims should be (cohort, product) = _dims_cohort.drop(time_letter)
[…]
