### 2025-12-11 75b7a8c
Remove duplicate is_computed property definition
Co-authored-by: JakobBD <44020564+JakobBD@users.noreply.github.com>
