### 2025-12-12 6b718d3
improve initial stock implementation
