+++ flodym/stocks.py
@@ -215,9 +215,7 @@ class DynamicStockModel(Stock):
                 raise ValueError("Cohort dimension size must be the same as time dimension size.")
             if c.letter in self.dims.letters or c.name in self.dims.names:
                 raise ValueError("Cohort dimension must not be part of the stock dimensions.")
-            self._dims_cohort = DimensionSet(
-                dim_list=[t, c] + list(self.dims.drop(self.time_letter, inplace=False).dim_list)
-            )
+            self._dims_cohort = t + c + self.dims.drop(self.time_letter, inplace=False)
         return self
 
     @model_validator(mode="after")
@@ -303,61 +301,96 @@ class DynamicStockModel(Stock):
             )
 
         inflow_by_cohort = StockArray(dims=self._dims_cohort, name=f"{self.name}_inflow_by_cohort")
-        # Manually set the slice for the initial_year time index
-        initial_year_idx = self.dims[self.time_letter].items.index(initial_year)
-        inflow_by_cohort.values[initial_year_idx, :, ...] = self._to_annual(initial_stock.values)
-        if np.any((inflow_by_cohort.values > 0) & (self.lifetime_model.sf == 0)):
-            raise ValueError(
-                "Initial stock is inconsistent with the lifetime model: survival function is zero for some cohort/inflow year combinations."
-            )
-        # Back-calculate the inflow from the initial stock by dividing by survival function at initial_year
-        # Then multiply by survival function at all times to get inflow_by_cohort[t, c]
-        inflow_reconstructed = inflow_by_cohort.values[initial_year_idx, :, ...] / np.where(
-            self.lifetime_model.sf[initial_year_idx, :, ...] == 0,
-            1,
-            self.lifetime_model.sf[initial_year_idx, :, ...]
-        )
-        # Set inflow_by_cohort for all time steps
-        for t_idx in range(self._n_t):
-            inflow_by_cohort.values[t_idx, :, ...] = inflow_reconstructed * self.lifetime_model.sf[t_idx, :, ...]
-        inflow_by_cohort.mark_set()  # Mark as set before passing to InflowByCohortDrivenDSM
-        self._initial_stock_dsm = InflowByCohortDrivenDSM(
+        inflow_by_cohort[{self.time_letter: initial_year}] = self._to_annual(initial_stock.values)
+        self._initial_stock_year = initial_year
+        isd = InflowByCohortDrivenDSM(
             dims=self.dims,
             cohort_dim=self.cohort_dim,
             lifetime_model=self.lifetime_model,
             inflow_by_cohort=inflow_by_cohort,
             name=f"{self.name}_ISD",
         )
-        self._initial_stock_year = initial_year
+        iiy = self._initial_year_index
+        # correction to account for the fact that some of the inflow will leave the stock in the same period
+        isd.inflow_by_cohort.values[iiy,:iiy+1,...] /= isd.lifetime_model.sf_conditional[iiy,iiy,:iiy+1,...]
+        self._initial_stock_dsm = isd
+
+    # TODO general:
+    # - np allclose in all
+    # - make work for inflow_by_cohort?
+    # + proper treatment of prescribed inflow and stock at initial year
+    #   (for inflow: only represents/replaces the newest cohort!)
+    # - add inflow_by_cohort to FlexibleDSM
+    # - move adapt_stock / adapt_inflow to respective subclasses
 
     @property
     def _initial_year_index(self):
         return self.dims[self.time_letter].items.index(self._initial_stock_year)
 
-    def _add_initial_stock_to_inflow(self):
-        if self._initial_stock_dsm is not None:
-            self._initial_stock_dsm.compute(stop_after="inflow")
-            if np.any(self.inflow.values[: self._initial_year_index + 1, ...] > 0):
-                raise ValueError(
[…]
