+++ flodym/lifetime_models.py
@@ -64,7 +64,9 @@ class LifetimeModel(PydanticBaseModel):
     Default is 1, meaning that the inflow is evaluated only once per time period.
     """
     _sf: np.ndarray = None
+    _conditional_sf: np.ndarray = None
     _pdf: np.ndarray = None
+    _conditional_pdf: np.ndarray = None
     _t: UnevenTimeDim = None
 
     @model_validator(mode="after")
@@ -108,6 +110,10 @@ class LifetimeModel(PydanticBaseModel):
         return (self._n_t,) + self.shape
 
     @property
+    def _shape_conditional(self):
+        return (self._n_t, self._n_t) + self.shape
+
+    @property
     def _shape_no_t(self):
         return tuple(list(self.shape)[1:])
 
@@ -115,15 +121,29 @@ class LifetimeModel(PydanticBaseModel):
     def sf(self):
         if self._sf is None:
             self._sf = np.zeros(self._shape_cohort)
-            self.compute_survival_factor()
+            self._compute_survival_factor()
         return self._sf
+
+    @property
+    def sf_conditional(self):
+        if self._sf_conditional is None:
+            self._sf_conditional = np.zeros(self._shape_conditional)
+            self._compute_sf_conditional()
+        return self._sf_conditional
 
     @property
     def pdf(self):
         if self._pdf is None:
             self._pdf = np.zeros(self._shape_cohort)
-            self.compute_outflow_pdf()
+            self._compute_outflow_pdf(self._pdf, self.sf)
         return self._pdf
+
+    @property
+    def pdf_conditional(self):
+        if self._pdf_conditional is None:
+            self._pdf_conditional = np.zeros(self._shape_conditional)
+            self._compute_outflow_pdf(self._pdf_conditional, self.sf)
+        return self._pdf_conditional
 
     def _tile(self, a: np.ndarray) -> np.ndarray:
         """tiles the input array a to the shape of the lifetime model, by adding non-time dimensions
@@ -143,7 +163,7 @@ class LifetimeModel(PydanticBaseModel):
         t = eta * self._t.bounds[m + 1] + (1 - eta) * self._t.bounds[m]
         return self._tile(self._t.bounds[m + 1 :] - t)
 
-    def compute_survival_factor(self):
+    def _compute_survival_factor(self):
         """Survival table self.sf(m,n) denotes the share of an inflow in year n (age-cohort) still
         present at the end of year m (after m-n years).
         The computation is self.sf(m,n) = ProbDist.sf(m-n), where ProbDist is the appropriate
@@ -162,13 +182,13 @@ class LifetimeModel(PydanticBaseModel):
         to save time.
         """
         self._check_prms_set()
-        quad_eta, quad_weights = self.get_quad_points_and_weights()
+        quad_eta, quad_weights = self._get_quad_points_and_weights()
         for m in range(0, self._n_t):  # cohort index
             for eta, weight in zip(list(quad_eta), list(quad_weights)):
                 t = self._remaining_ages(m, eta)
                 self._sf[m::, m, ...] += weight * self._survival_by_year_id(t, m)
 
-    def get_quad_points_and_weights(self):
+    def _get_quad_points_and_weights(self):
         """Returns the quadrature points and weights for the inflow time periods."""
         if self.n_pts_per_interval > 10:
             raise ValueError("quad_order must be between 0 and 9.")
@@ -200,14 +220,19 @@ class LifetimeModel(PydanticBaseModel):
             prm_out[...] = prm_in
[…]
