### 2025-12-12 1b37e90
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
