### 2025-12-12 6205f74
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
