+++ tests/test_stocks.py
@@ -178,13 +178,13 @@ def test_inflow_by_cohort_driven_dsm():
             Dimension(name="product", letter="p", items=["A", "B"], dtype=str),
         ]
     )
-    
+
     # Create cohort dimension
     cohort_dim = Dimension(name="cohort", letter="c", items=time_items, dtype=int)
-    
+
     # Create lifetime model
     lifetime_model = LogNormalLifetime(dims=dims_test, time_letter="t", mean=5, std=2)
-    
+
     # Create InflowByCohortDrivenDSM
     dsm = InflowByCohortDrivenDSM(
         dims=dims_test,
@@ -193,23 +193,23 @@ def test_inflow_by_cohort_driven_dsm():
         time_letter="t",
         name="test_dsm",
     )
-    
+
     # Set inflow by cohort - diagonal pattern (inflow only occurs in cohort year)
     dsm.inflow_by_cohort.values[...] = 0.0
     for i in range(len(time_items)):
         dsm.inflow_by_cohort.values[i, i, :] = 1.0  # 1 unit per product per year
-    
+
     # Compute the DSM
     dsm.compute()
-    
+
     # Check that inflow was computed correctly (diagonal extraction)
     assert dsm.inflow.is_set
     assert dsm.stock.is_set
     assert dsm.outflow.is_set
-    
+
     # Verify inflow values are approximately 1.0 at each time step
     assert np.allclose(dsm.inflow.values, 1.0)
-    
+
     # Verify stock balance
     dsm.check_stock_balance()
 
@@ -224,13 +224,13 @@ def test_inflow_driven_dsm_with_initial_stock():
             Dimension(name="product", letter="p", items=["A", "B"], dtype=str),
         ]
     )
-    
+
     # Create cohort dimension (required for initial stock functionality)
     cohort_dim = Dimension(name="cohort", letter="c", items=time_items, dtype=int)
-    
+
     # Create lifetime model
     lifetime_model = LogNormalLifetime(dims=dims_test, time_letter="t", mean=8, std=3)
-    
+
     # Create InflowDrivenDSM with cohort_dim
     dsm = InflowDrivenDSM(
         dims=dims_test,
@@ -239,7 +239,7 @@ def test_inflow_driven_dsm_with_initial_stock():
         time_letter="t",
         name="test_dsm_initial",
     )
-    
+
     # Set initial stock at year 2005
     initial_year = 2005
     # Initial stock dims should be (cohort, product) = _dims_cohort.drop(time_letter)
@@ -251,30 +251,30 @@ def test_inflow_driven_dsm_with_initial_stock():
     # Set stock from various cohorts
     for i in range(5):  # 5 historical cohorts
         initial_stock.values[i, :] = 10.0 - i  # Decreasing stock from older cohorts
-    
+
     dsm.set_initial_stock(initial_stock, initial_year)
-    
+
     # Set inflow for years AFTER initial year only
     # Inflow before and at initial year will be computed from initial stock
     dsm.inflow.values[...] = 0.0
     initial_year_idx = time_items.index(initial_year)
     dsm.inflow.values[initial_year_idx + 1 :, :] = 2.0  # 2 units per year after initial
     dsm.inflow.mark_set()  # Mark inflow as set
-    
+
     # Compute
     dsm.compute()
-    
+
     # Verify that stock is set
     assert dsm.stock.is_set
     assert dsm.inflow.is_set
     assert dsm.outflow.is_set
-    
+
     # Stock at initial year should be non-zero
     assert dsm.stock.values[initial_year_idx, :].sum() > 0
-    
+
     # Stock before initial year should also be non-zero (reconstructed from initial stock)
[…]
