### 2026-01-06 364db62
add total_size property to DimensionSet
