### 2026-01-10 785528f
add numpy-like helpers
