### 2026-01-10 bf062dc
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
