### 2026-01-10 0ce3542
add tests to convenience functions
