### 2026-01-12 82eab20
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
