### 2026-01-13 d385be6
refactor: simplify zero-check conditions in InflowDrivenDSM and StockDrivenDSM
