### 2026-01-13 fe41390
revert unecessary changes, improve code and documentation
