### 2026-01-13 d0571b5
remove tests for removed functions, increase test scrutiny for remaining functions
