+++ tests/test_flodym_arrays.py
@@ -226,77 +226,189 @@ def test_to_df():
         df.loc[("Earth", 2000, "mouse"), "value"]
 
 
-def test_flodym_array_zeros_helper():
-    zeros = FlodymArray.zeros(dims, name="zeros")
-    assert zeros.name == "zeros"
-    assert zeros.dims == dims
-    assert zeros.values.shape == dims.shape
-    assert np.count_nonzero(zeros.values) == 0
+class TestFlodymArrayFull:
+    """Tests for FlodymArray.full() class method with various fill value types."""
+
+    def test_full_with_int(self):
+        filled = FlodymArray.full(dims, fill_value=3, name="int_filled")
+        assert filled.name == "int_filled"
+        assert filled.dims == dims
+        assert filled.dims is not dims  # ensure dims is copied
+        assert np.all(filled.values == 3)
+        assert filled.values.dtype == np.int_
+
+    def test_full_with_float(self):
+        filled = FlodymArray.full(dims, fill_value=3.5, name="float_filled")
+        assert filled.name == "float_filled"
+        assert filled.dims == dims
+        assert np.all(filled.values == 3.5)
+        assert filled.values.dtype == np.float64
+
+    def test_full_with_np_array(self):
+        """Test full with a numpy array that broadcasts across dimensions."""
+        # Create a 1D array that broadcasts along the time dimension
+        time_values = np.array([1.0, 2.0, 3.0])  # shape (3,) for 3 time steps
+        filled = FlodymArray.full(dims, fill_value=time_values, name="np_array_filled")
+        assert filled.name == "np_array_filled"
+        assert filled.values.shape == dims.shape  # (4, 3)
+        # Each row should have [1.0, 2.0, 3.0]
+        assert_array_equal(filled.values[0], time_values)
+        assert_array_equal(filled.values[1], time_values)
+        assert_array_equal(filled.values[3], time_values)
+
+    def test_full_with_2d_np_array(self):
+        """Test full with a 2D numpy array matching the full shape."""
+        fill_array = np.arange(12).reshape(4, 3).astype(np.float64)
+        filled = FlodymArray.full(dims, fill_value=fill_array, name="2d_array_filled")
+        assert filled.name == "2d_array_filled"
+        assert_array_equal(filled.values, fill_array)
+        assert filled.values.dtype == np.float64
+
+    def test_full_dims_not_modified(self):
+        """Ensure the original dims is not modified when creating a full array."""
+        original_dims = DimensionSet(dim_list=base_dim_list)
+        filled = FlodymArray.full(original_dims, fill_value=1.0)
+        # Modify the filled array's dims
+        filled.dims.dim_list[0] = animals
+        # Original should be unchanged
+        assert original_dims[0] == places
 
 
-def test_flodym_array_ones_helper():
-    ones = FlodymArray.ones(dims, name="ones")
-    assert ones.name == "ones"
-    assert ones.dims == dims
-    assert ones.values.shape == dims.shape
-    assert np.all(ones.values == 1)
+class TestFlodymArrayFullLike:
+    """Tests for FlodymArray.full_like() class method with various fill value types."""
+
+    def test_full_like_with_int(self):
+        template = FlodymArray.full(dims, fill_value=2.0, name="template")
+        filled_like = FlodymArray.full_like(template, fill_value=5, name="int_filled")
+        assert filled_like.name == "int_filled"
+        assert filled_like.dims == template.dims
+        assert filled_like.dims is not template.dims  # ensure dims is copied
+        assert np.all(filled_like.values == 5)
+        assert filled_like.values.dtype == np.int_
+
+    def test_full_like_with_float(self):
+        template = FlodymArray.full(dims, fill_value=2.0, name="template")
[…]
