### 2026-01-13 90e69f7
ensure get_subset() returns a propper copy of the DimensionSet.dim_list
