+++ flodym/stocks.py
@@ -267,7 +267,7 @@ class StockDrivenDSM(DynamicStockModel):
 
     def _check_needed_arrays(self):
         super()._check_needed_arrays()
-        if np.allclose(self.inflow.values, np.zeros(self.shape)):
+        if np.allclose(self.stock.values, np.zeros(self.shape)):
             logging.warning("Stock is zero. This will lead to a zero inflow and outflow.")
 
     def compute(self):
