### 2026-01-13 6f3ac28
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
