### 2026-01-13 0318c3b
refactor: update FlodymArray to use Number type
