+++ flodym/flodym_arrays.py
@@ -11,6 +11,7 @@ import pandas as pd
 from pydantic import BaseModel as PydanticBaseModel, ConfigDict, model_validator
 from typing import Optional, Union, Callable
 from copy import copy
+from numbers import Number
 
 from .processes import Process
 from .dimensions import DimensionSet, Dimension
@@ -57,7 +58,7 @@ class FlodymArray(PydanticBaseModel):
 
     dims: DimensionSet
     """Dimensions of the FlodymArray."""
-    values: Optional[Union[np.ndarray, np.generic]] = None
+    values: Optional[Union[np.ndarray, Number]] = None
     """Values of the FlodymArray. Must have the same shape as the dimensions of the FlodymArray. If None, an array of zeros is created."""
     name: Optional[str] = "unnamed"
     """Name of the FlodymArray."""
@@ -76,11 +77,11 @@ class FlodymArray(PydanticBaseModel):
         return self
 
     def _check_value_format(self):
-        if not isinstance(self.values, (np.ndarray, np.generic)):
-            raise ValueError("Values must be a numpy array or numpy generic.")
+        if not isinstance(self.values, (np.ndarray, Number)):
+            raise ValueError("Values must be a numpy array or Number.")
         if self.dims.ndim > 0 and not isinstance(self.values, np.ndarray):
             raise ValueError("Values must be a numpy array, except for 0-dimensional arrays.")
-        elif self.dims.ndim == 0 and isinstance(self.values, np.generic):
+        elif self.dims.ndim == 0 and isinstance(self.values, Number):
             self.values = np.array(self.values)
 
         if self.values.shape != self.dims.shape:
@@ -112,14 +113,14 @@ class FlodymArray(PydanticBaseModel):
     def full(
         cls,
         dims: DimensionSet,
-        fill_value: Union[int, float, np.number, np.ndarray],
+        fill_value: Union[Number, np.ndarray],
         **kwargs,
     ) -> "FlodymArray":
         """Create a FlodymArray filled with a constant value for the provided dimensions.
 
         Parameters:
             dims (DimensionSet): DimensionSet defining the dimensions of the FlodymArray.
-            fill_value (Union[int, float, np.number, np.ndarray]): Value to fill the array with.
+            fill_value (Union[Number, np.ndarray]): Value to fill the array with.
                 Can be a scalar or an array that is broadcastable to the shape of dims.
             **kwargs: Additional keyword arguments passed to the FlodymArray constructor
                 (e.g., name).
@@ -133,7 +134,7 @@ class FlodymArray(PydanticBaseModel):
     def full_like(
         cls,
         other: "FlodymArray",
-        fill_value: Union[int, float, np.number, np.ndarray],
+        fill_value: Union[Number, np.ndarray],
         dtype: Optional[Union[type, np.dtype]] = None,
         **kwargs,
     ) -> "FlodymArray":
@@ -141,7 +142,7 @@ class FlodymArray(PydanticBaseModel):
 
         Parameters:
             other (FlodymArray): FlodymArray whose dimensions will be used for the new array.
-            fill_value (Union[int, float, np.number, np.ndarray]): Value to fill the array with.
+            fill_value (Union[Number, np.ndarray]): Value to fill the array with.
                 Can be a scalar or an array that is broadcastable to the shape of other.
             dtype (Optional[Union[type, np.dtype]], optional): Data type of the new array.
                 If None, the data type of fill_value is used. Defaults to None.
@@ -163,13 +164,13 @@ class FlodymArray(PydanticBaseModel):
     @classmethod
     def scalar(
         cls,
-        value: Union[int, float, np.number],
+        value: Number,
         **kwargs,
     ) -> "FlodymArray":
         """Create a scalar (zero-dimensional) FlodymArray.
 
         Parameters:
[…]
