### 2026-01-15 8faba2e
add and use DimensionSet copy method
