### 2026-01-16 ab2bb76
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
