+++ flodym/flodym_arrays.py
@@ -478,9 +478,7 @@ class FlodymArray(PydanticBaseModel):
         Returns:
             FlodymArray: A new FlodymArray object with copied values and dimensions.
         """
-        return self.model_copy(
-            update={"dims": self.dims.copy(), "values": self.values.copy()}
-        )
+        return self.model_copy(update={"dims": self.dims.copy(), "values": self.values.copy()})
 
     def abs(self, inplace: bool = False):
         return self.apply(np.abs, inplace=inplace)
