### 2026-01-16 7eb873d
replace dimension copy with deepcopy, add tests
