### 2026-01-16 712b9e1
refactor: copy doesn't deepcopy
