### 2026-01-16 88627de
copy dim_list upon creation of new DimensionSet objects, to ensure that inplace operations do not affect other DimensionSet objects
