+++ flodym/dimensions.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
-from copy import copy, deepcopy
+from copy import copy
 from pydantic import BaseModel as PydanticBaseModel, Field, AliasChoices, model_validator
 from typing import Dict, Iterator, Optional
 import numpy as np
@@ -153,11 +153,6 @@ class DimensionSet(PydanticBaseModel):
     dim_list: list[Dimension]
     """A list of Dimension objects defining the set"""
 
-    @classmethod
-    def empty(cls) -> "DimensionSet":
-        """Return an empty DimensionSet."""
-        return cls(dim_list=[])
-
     @model_validator(mode="after")
     def no_repeated_dimensions(self):
         """Check that all dimensions have unique letters."""
@@ -165,6 +160,18 @@ class DimensionSet(PydanticBaseModel):
         if len(letters) != len(set(letters)):
             raise ValueError("Dimensions must have unique letters in DimensionSet.")
         return self
+
+    @model_validator(mode="after")
+    def copy_dim_list(self):
+        """Ensure each DimensionSet has its own copy of the dim_list to avoid inplace operations
+        affecting other DimensionSets."""
+        self.dim_list = copy(self.dim_list)
+        return self
+
+    @classmethod
+    def empty(cls) -> "DimensionSet":
+        """Return an empty DimensionSet."""
+        return cls(dim_list=[])
 
     @property
     def _full_mapping(self) -> Dict[str, Dimension]:
@@ -223,25 +230,40 @@ class DimensionSet(PydanticBaseModel):
 
     def copy(self) -> "DimensionSet":
         """Return a deep copy of the DimensionSet."""
-        return self.model_copy(update={"dim_list": deepcopy(self.dim_list)})
+        return self.model_copy(update={"dim_list": copy(self.dim_list)})
 
     def get_subset(self, dims: tuple = None) -> "DimensionSet":
         """Selects :py:class:`Dimension` objects from the object attribute dim_list,
         according to the dims passed, which can be either letters or names.
         Returns a copy if dims are not given.
+
+        Args:
+            dims (tuple, optional): A tuple of dimension letters or names to select. Defaults to None.
         """
         subset = self.model_copy()
         if dims is not None:
             subset.dim_list = [self._full_mapping[dim_key] for dim_key in dims]
         return subset
 
-    def expand_by(self, added_dims: list[Dimension]) -> "DimensionSet":
-        """Expands the DimensionSet by adding new dimensions to it."""
+    def expand_by(self, added_dims: list[Dimension], inplace: bool = False) -> "DimensionSet":
+        """Expands the DimensionSet by adding new dimensions to it.
+
+        Args:
+            added_dims (list[Dimension]): A list of Dimension objects to add
+            inplace (bool, optional): If True, the operation is performed in place, otherwise a new DimensionSet is returned. Defaults to False.
+
+        Returns:
+            None if inplace=True, otherwise a new DimensionSet with the new dimensions added
+        """
         if not all([dim.letter not in self.letters for dim in added_dims]):
             raise ValueError(
                 "DimensionSet already contains one or more of the dimensions to be added."
             )
-        return DimensionSet(dim_list=self.dim_list + added_dims)
+        if inplace:
+            self.dim_list.extend(added_dims)
+            return
+        else:
+            return DimensionSet(dim_list=self.dim_list + added_dims)
 
     extend = expand_by
 
@@ -274,9 +296,11 @@ class DimensionSet(PydanticBaseModel):
 
     def prepend(self, new_dim: Dimension, inplace: bool = False) -> Optional["DimensionSet"]:
         """Add a new dimension to the set at the beginning.
+
         Args:
             new_dim (Dimension): The new dimension to add
[…]
