### 2026-01-16 d271516
Revert "refactor: copy doesn't deepcopy"
This reverts commit 712b9e17e4fb21f405e2544e6fc0ed2a20be81c9.
