### 2026-01-16 87eea02
first solving attempt
