### 2026-01-16 5d6d00e
revert solving attempt
