### 2026-01-16 b851548
fix numpy indexing error in SubArrayHandler
