+++ tests/test_flodym_arrays.py
@@ -144,14 +144,6 @@ def test_maths():
     assert_array_almost_equal(divided_flipped.values[:, :, 1], values / (animal_values[:, :, 1]))
 
 
-def test_get_item():
-    cats_on_the_moon = space_animals["Moon"]["cat"]
-    assert isinstance(cats_on_the_moon, FlodymArray)
-    assert_array_almost_equal(cats_on_the_moon.values, space_animals.values[2, :, 0])
-    # note that this does not work for the time dimension (not strings)
-    # and also assumes that no item appears in more than one dimension
-
-
 def test_sub_array_handler():
     space_cat = space_animals["cat"]  # space cat from str
     another_space_cat = space_animals[{"a": "cat"}]  # space cat from dict
@@ -191,83 +183,69 @@ def test_dimension_subsets():
     with pytest.raises(ValueError):
         historic_space_animals[{"h": time}]  # index is not a subset
 
+class TestFlodymArrayIndexing:
+    """Tests for getitem and setitem, including with advanced indexing separated by slices."""
 
-def test_multiple_dimension_subsets():
-    """Test slicing with multiple Dimension objects at once.
-
-    This reproduces an issue where slicing with two Dimension objects causes
-    a shape mismatch error due to NumPy's advanced indexing broadcasting behavior.
-    """
-    # Create a larger array to make the issue more apparent
-    regions = Dimension(name="region", letter="r", items=["EUR", "USA", "CHN"])
-    years = Dimension(name="time", letter="t", items=list(range(2020, 2030)))  # 10 years
-    materials = Dimension(name="material", letter="m", items=["concrete", "steel", "wood"])
-    applications = Dimension(
-        name="application", letter="a", items=["A1", "A2", "A3", "A4", "A5", "A6"]
-    )
-
-    full_dims = DimensionSet(dim_list=[regions, years, materials, applications])
+    # 80 years
+    years = Dimension(name="year", letter="t", items=list(range(2020, 2100)))
+    # 12 regions
+    regions = Dimension(name="region", letter="r", items=[f"R{i}" for i in range(1, 13)])
+    # 4 sectors
+    sectors = Dimension(name="sector", letter="s", items=["S1", "S2", "S3", "S4"])
+    # 7 products
+    products = Dimension(name="product", letter="p", items=["P1", "P2", "P3", "P4", "P5", "P6", "P7"])
+    # 2 materials
+    materials = Dimension(name="material", letter="m", items=["M1", "M2"])
+    full_dims = DimensionSet(dim_list=[years, regions, sectors, products, materials])
     arr = FlodymArray(
         dims=full_dims,
         values=np.arange(np.prod(full_dims.shape)).reshape(full_dims.shape).astype(float),
     )
+    # 5 regions in subset
+    subset_regions = Dimension(name="sub_regions", letter="x", items=["R1", "R2", "R3", "R4", "R5"])
+    # 3 products in subset
+    subset_products = Dimension(name="sub_products", letter="y", items=["P1", "P2", "P3"])
+    # mask includes one material, subset regions and subset products
+    mask = {"m": "M1", "r": subset_regions, "p": subset_products}
 
-    # Create subset dimensions
-    subset_years = Dimension(
-        name="subset_years", letter="y", items=[2025, 2026, 2027, 2028, 2029]
-    )  # 5 years
-    subset_apps = Dimension(
-        name="subset_apps", letter="b", items=["A1", "A2", "A3", "A4"]
-    )  # 4 apps
+    def test_get_item(self):
+        cats_on_the_moon = space_animals["Moon"]["cat"]
+        assert isinstance(cats_on_the_moon, FlodymArray)
[…]
