### 2026-01-17 3bdca49
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
