### 2026-01-19 bff5f4c
fix further indexing bug
