### 2026-01-19 df6e7db
add some comments to subarrayhandler
