### 2026-01-19 c4988fd
remove unnecessary main guard from test files
