+++ paper/paper.bib
@@ -179,3 +179,129 @@ year = {2025},
 url = {https://transience-eu-mfa.readthedocs.io}
 }
 
+
+@misc{AutumnSchool2025Website,
+	author = {Mutel, Chris},
+	title = {{DdS Autumn School 2025}},
+	year = {2025},
+	month = jul,
+	note = {[Online; accessed 19. Jan. 2026]},
+	url = {https://www.d-d-s.ch/schools/nov-25}
+}
+@misc{ISIE_SEM_2024,
+	title = {{ISIE-SEM Summer School - International Society for Industrial Ecology - ISIE}},
+	year = {2024},
+	month = jan,
+	note = {[Online; accessed 19. Jan. 2026]},
+	url = {https://is4ie.org/events/event/isie-sem-summer-school}
+}
+
+@misc{AutumnSchool2025Github,
+	title = {{Schools-2025-November-Switzerland}},
+	journal = {GitHub},
+	year = {2026},
+	month = jan,
+	note = {[Online; accessed 19. Jan. 2026]},
+	url = {https://github.com/Depart-de-Sentier/Schools-2025-November-Switzerland}
+}
+
+@misc{PAW_MFA_LCA_2025,
+	title = {{PAW{$\_$}MFA{$\_$}LCA{$\_$}2025}},
+	journal = {GitHub},
+	year = {2026},
+	month = jan,
+	note = {[Online; accessed 19. Jan. 2026]},
+	url = {https://github.com/isabelapi/PAW_MFA_LCA_2025}
+}
+
+@misc{DdS_REFLOC,
+	title = {{DdS{$\_$}REFLOC}},
+	journal = {GitHub},
+	year = {2026},
+	month = jan,
+	note = {[Online; accessed 19. Jan. 2026]},
+	url = {https://github.com/MeYiwen/DdS_REFLOC}
+}
+
+@misc{PVProject,
+	title = {{PVProject}},
+	journal = {GitHub},
+	year = {2026},
+	month = jan,
+	note = {[Online; accessed 19. Jan. 2026]},
+	url = {https://github.com/Tanima-Sharma/PVProject}
+}
+
+@misc{dds2025manure,
+	title = {{dds2025manure}},
+	journal = {GitHub},
+	year = {2026},
+	month = jan,
+	note = {[Online; accessed 19. Jan. 2026]},
+	url = {https://github.com/gergosuto/dds2025manure}
+}
+
+
+@article{Muller2014Feb,
+	author = {M{\ifmmode\ddot{u}\else\"{u}\fi}ller, Esther and Hilty, Lorenz M. and Widmer, Rolf and Schluep, Mathias and Faulstich, Martin},
+	title = {{Modeling Metal Stocks and Flows: A Review of Dynamic Material Flow Analysis Methods}},
+	journal = {Environ. Sci. Technol.},
+	volume = {48},
+	number = {4},
+	pages = {2102--2113},
+	year = {2014},
+	month = feb,
+	issn = {0013-936X},
+	publisher = {American Chemical Society},
+	doi = {10.1021/es403506a}
+}
+
+@incollection{Bringezu2018Jan,
+	author = {Bringezu, Stefan and Moriguchi, Yuichi},
+	title = {{Material flow analysis}},
+	booktitle = {{Green Accounting}},
+	journal = {Taylor {\&} Francis},
+	pages = {149--166},
+	year = {2018},
+	month = jan,
+	publisher = {Routledge},
+	address = {New York, NY, USA},
+	doi = {10.4324/9781315197715-6}
+}
+
+@article{Graedel2019Nov,
+	author = {Graedel, Thomas E.},
+	title = {{Material Flow Analysis from Origin to Evolution}},
+	journal = {Environ. Sci. Technol.},
+	volume = {53},
+	number = {21},
[…]
