+++ paper/paper.bib
@@ -179,68 +179,62 @@ year = {2025},
 url = {https://transience-eu-mfa.readthedocs.io}
 }
 
-
-@misc{AutumnSchool2025Website,
-	author = {Mutel, Chris},
-	title = {{DdS Autumn School 2025}},
-	year = {2025},
-	month = jul,
-	note = {[Online; accessed 19. Jan. 2026]},
-	url = {https://www.d-d-s.ch/schools/nov-25}
-}
-@misc{ISIE_SEM_2024,
-	title = {{ISIE-SEM Summer School - International Society for Industrial Ecology - ISIE}},
-	year = {2024},
-	month = jan,
-	note = {[Online; accessed 19. Jan. 2026]},
-	url = {https://is4ie.org/events/event/isie-sem-summer-school}
+@online{AutumnSchool2025Website,
+author = {Mutel, Chris},
+title = {DdS Autumn School 2025},
+year = {2025},
+month = jul,
+url = {https://www.d-d-s.ch/schools/nov-25},
+urldate = {2026-01-19}
 }
 
-@misc{AutumnSchool2025Github,
-	title = {{Schools-2025-November-Switzerland}},
-	journal = {GitHub},
-	year = {2026},
-	month = jan,
-	note = {[Online; accessed 19. Jan. 2026]},
-	url = {https://github.com/Depart-de-Sentier/Schools-2025-November-Switzerland}
+@online{ISIE_SEM_2024,
+title = {ISIE-SEM Summer School},
+year = {2024},
+month = jan,
+url = {https://is4ie.org/events/event/isie-sem-summer-school},
+urldate = {2026-01-19}
 }
 
-@misc{PAW_MFA_LCA_2025,
-	title = {{PAW{$\_$}MFA{$\_$}LCA{$\_$}2025}},
-	journal = {GitHub},
-	year = {2026},
-	month = jan,
-	note = {[Online; accessed 19. Jan. 2026]},
-	url = {https://github.com/isabelapi/PAW_MFA_LCA_2025}
+@online{AutumnSchool2025Github,
+title = {Schools-2025-November-Switzerland},
+year = {2026},
+month = jan,
+url = {https://github.com/Depart-de-Sentier/Schools-2025-November-Switzerland},
+urldate = {2026-01-19}
 }
 
-@misc{DdS_REFLOC,
-	title = {{DdS{$\_$}REFLOC}},
-	journal = {GitHub},
-	year = {2026},
-	month = jan,
-	note = {[Online; accessed 19. Jan. 2026]},
-	url = {https://github.com/MeYiwen/DdS_REFLOC}
+@online{PAW_MFA_LCA_2025,
+title = {PAW_MFA_LCA_2025},
+year = {2026},
+month = jan,
+url = {https://github.com/isabelapi/PAW_MFA_LCA_2025},
+urldate = {2026-01-19}
 }
 
-@misc{PVProject,
-	title = {{PVProject}},
-	journal = {GitHub},
-	year = {2026},
-	month = jan,
-	note = {[Online; accessed 19. Jan. 2026]},
-	url = {https://github.com/Tanima-Sharma/PVProject}
+@online{DdS_REFLOC,
+title = {DdS_REFLOC},
+year = {2026},
+month = jan,
+url = {https://github.com/MeYiwen/DdS_REFLOC},
+urldate = {2026-01-19}
 }
 
-@misc{dds2025manure,
-	title = {{dds2025manure}},
-	journal = {GitHub},
-	year = {2026},
-	month = jan,
-	note = {[Online; accessed 19. Jan. 2026]},
-	url = {https://github.com/gergosuto/dds2025manure}
+@online{PVProject,
+title = {PVProject},
[…]
