### 2026-01-19 #8685 @JakobBD
@kyleniemeyer

I can very well imagine that generative AI poses a challenge to JOSS.
I have updated the paper accordingly.

This paper has been in the queue for six months and was seemingly one of the next in line.
I hope queuing doesn't start from zero now?

Thanks for your support!
