### 2026-01-20 #8685 @kyleniemeyer
@JakobBD thank you—we will evaluate based on our current scope requirements. The backlog/queue remains the same, although some papers may be removed for not meeting our new criteria.
