### 2026-02-23 #10105 @editorialbot
[REVIEW]: flodym: A Python package for dynamic material flow analysis
**Submitting author:** @JakobBD (<a href="http://orcid.org/0000-0001-8961-5340">Jakob Dürrwächter</a>)
**Repository:** https://github.com/pik-piam/flodym
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.7.0
