### 2026-03-05 #10105 @TimoDiepers
@JakobBD Some comments on the paper:
- Many of the points from the „State of the field“ section describe the features / design of flodym. I’d suggest to move these points (examples and descriptions of the simplifications flodym enables) to the software design section, and rather focus the state of the field section on the description (including potential downsides) of existing software. This would also resolve some of the redundancy that there currently is between the two sections.
- Unify how courses are cited. For example, compare „ISIE SEM summer school“ and „Schools-2025-november-switzerland“. For the latter, you might also consider updating the link to point to the actual event (https://www.d-d-s.ch/schools/nov-25/)
- Language-related:
    - "MFA provides quantitative insights _**into**_ how materials are extracted"
    - "...not reached publication yet, but **_its_** external user base is growing rapidly."
    - Product Names: Github -> GitHub, Github actions -> GitHub Actions, VSCode -> VS Code
    - "flodym" is sometimes written in italics, sometimes not.
