### 2026-03-05 #10105 @JakobBD
Thanks @TimoDiepers and @michaelweinold for reviewing!!
And thanks already for the helpful remarks.

I'll wait until both your reviews are ready (could you please give me a short note when they are?) and then address all your suggestions!
