### 2026-03-05 #10105 @michaelweinold
# Comments on the Publication:

## General Comments

The article is on the longer side for JOSS manuscripts. It should be shortened possible through the removal of:

- The "Software design" section, which currently includes mostly subjective and unhelpful phrasing, such as "Providing a coherent, as clean as possible API (...)" and "What’s [sic!] more, flodym features different levels of integration (...).
- Large parts of the "Research impact statement", which starts with the unhelpful statement that "flodym is a new tool and therefore applications have not reached publication yet". The list of "teaching events" should also include only past events. The mention of the number of GitHub stars, which is both [easily manipulated](https://medium.com/@vanshagrawal561/i-bought-fake-github-stars-so-you-dont-have-to-88f12c8eccb6) and (hopefully) subject to changes in the future should also be removed.

## Specific Comments

- Ensure you consistently name/stylize software packages. You write _flodym_, but also eg. pandas ([consistent with the project](https://pandas.pydata.org/about/)) and numpy ([inconsistent with the project](https://numpy.org/about/)).

## Comments by Section

### Summary

In general, this section is well written - but in "(...) including sanity checks for the system", what are those sanity checks - are they for non-negative flows, etc.?

### Statement of need

I recommend removing the sentence including the publication count "Scopus lists 4260 publications (...)", since this number is meaningless on its own - and no year for the search was given.

The unnecessary brackets in citations of the second paragraph needs to be removed: Instead of "(...) for example, (Müller et al., 2014), (Bringezu & Moriguchi, 2018), (...)", it should be "(...) for example, Müller et al., 2014, Bringezu & Moriguchi, 2018, (...)".

In "This makes general and easily accessible MFA tools (...)" what would be a "general MFA tool"?

The abbreviation "GUI" is introduced but never defined.

The "L" in "(...) tools like Life cycle (...)" should not be capitalized.

"(...) where data can be exchanged flexibly through a coded solution." sounds a bit awkward.

### State of the field

While you mention a few other tools at the end of the section, you really only compare your tool to ODYM here... I would therefore suggest you rename this section to something including "ODYM".

Ensure you consistently use punctuation in enumerations - currently only some of your bullet points end with a full stop.

In your code examples, for instance, `waste = np.einsum('trp,pw->trw', end_of_life_products, waste_share)` - what does this do? You cannot assume the reader knows what `trp`, etc. are without having used ODYM. You might removing un-interesting paragraphs on plotting functionality while expanding more on the (syntax)-benefits of flowdym.

### Acknowledgements

The authors mix Third-Person Objective ("We acknowledge...") and First-Person Active voice ("Thank you...").

### References

References are a mess and need to be unified:

- Some items give metadata description (`[Computer software]`), others leave the reader guessing from the URL (`dds2025manure. (2026, January). https://github.com/gergosuto/dds2025manure`).
- Some items incorrectly use the _in proceedings_ descriptor "In", even though the reference is clearly to the homepage of an event ("Brightcon 2025, hackathon & courses in Grenoble and online. (2026). In Indico. (...)"
- Plotly Technologies Inc. is apparently so important it is mentioned twice in the same reference.
- Capitalization of reference titles must be fixed in many places. For example: "MFA using r" (the "R" is presumably the programming language)
