### 2026-03-05 #10105 @michaelweinold
@marlinarnz & @JakobBD, I have concluded my initial review - no comments on the software itself, where I could successfully install and follow though all examples... only a few notes on the paper.
