+++ paper/paper.md
@@ -30,88 +30,98 @@ bibliography: paper.bib
 
 Material flow analysis (MFA) is a core method in industrial ecology that tracks time-dependent material flows within a system, such as a national economy, across all life cycle stages. It also accounts for material accumulation in stocks, including materials embodied in products, assets, and infrastructure at a given time. MFA supports resource management, environmental impact assessment, and the evaluation of circular economy strategies, informing policy advice, urban and regional planning, and sustainable product design.
 
-*flodym* (Flexible Open Dynamic Material Systems Model) is a library of objects and functions needed to build dynamic MFA. Mathematically, a large part of MFA translates to operations between multi-dimensional arrays.
-*flodym* implements the `FlodymArray` class, which internally manages operations of one or several such arrays. Objects representing flows, stocks, and parameters all inherit from this class. Stocks include lifetime models for dynamic stock modelling, i.e. for calculating the relation of material flows entering a stock and the mass and age structure of that stock over time. The whole MFA system is realized with an abstract parent class including sanity checks for the system, that users can implement a subclass of. *flodym* includes functionality for efficient read-in and export via pandas [@pandas1], [@pandas2], as well as visualization routines.
+flodym (Flexible Open Dynamic Material Systems Model) is a library of objects and functions needed to build dynamic MFA. Mathematically, a large part of MFA translates to operations between multi-dimensional arrays.
+flodym implements the `FlodymArray` class, which internally manages operations of one or several such arrays. Objects representing flows, stocks, and parameters all inherit from this class. Stocks include lifetime models for dynamic stock modelling, i.e. for calculating the relation of material flows entering a stock and the mass and age structure of that stock over time. The whole MFA system is realized with an abstract parent class, which users can implement a subclass of. The parent class includes checking routines for mass balance consistency and non-negative flows. flodym includes functionality for efficient read-in and export via pandas [@pandas1], [@pandas2], as well as visualization routines.
 
-*flodym* is based on the concepts of the Open Dynamic Material Systems Model (ODYM) [@odym]. It can be seen as a re-implementation with vastly expanded functionality and improved structuring. As a result, *flodym* enables users to write customized, flexible MFAs, built for maintainability and future expansion.
+flodym is based on the concepts of the Open Dynamic Material Systems Model (ODYM) [@odym]. It can be seen as a re-implementation with vastly expanded functionality and improved structuring. As a result, flodym enables users to write customized, flexible MFAs, built for maintainability and future expansion.
 
 # Statement of need
 
-MFA provides quantitative insights how materials are extracted, transformed, used, and retained within socio-economic systems over time. In academia, it enables research on resource efficiency, environmental impacts, and sustainability transitions. In industry, it informs decisions on material sourcing, product lifetimes, and circular economy strategies. For policy-makers and planners, MFA provides an evidence base for resource governance, waste and demand forecasting, and the design of interventions.
+MFA provides quantitative insights into how materials are extracted, transformed, used, and retained within socio-economic systems over time. In academia, it enables research on resource efficiency, environmental impacts, and sustainability transitions. In industry, it informs decisions on material sourcing, product lifetimes, and circular economy strategies. For policy-makers and planners, MFA provides an evidence base for resource governance, waste and demand forecasting, and the design of interventions.
 
-MFA is therefore a widespread method in industrial ecology and related fields. Scopus lists 4260 publications including the exact term "Material flow analysis" in title or abstract. Review papers include, for example, [@Muller2014Feb], [@Bringezu2018Jan], [@Graedel2019Nov], and [@Streeck2023Apr].
[…]
