+++ paper/paper.bib
@@ -48,7 +48,7 @@ year = {2015}
 }
 
 
-@manual{pymfa2.1,
+@software{pymfa2.1,
 author = {Thiébaud, Esther and Alexandru, Carol and Badat, Rolf and Kohler, Dimitri and Hilty, Lorenz},
 title = {pymfa2: Ein Werkzeug zur Analyse von Materialflüssen in Python 2.7},
 version = {2.1},
@@ -57,9 +57,9 @@ url = {https://bitbucket.org/Xeelk/pymfa2/src/master/}
 }
 
 
-@manual{PMFA,
+@software{PMFA,
 author = {Kawecki-Wenger, Delphine},
-title = {PMFA: Base functions for performing a Probabilistic MFA using R},
+title = {PMFA: Base functions for performing a Probabilistic MFA using {R}},
 url = {https://github.com/empa-tsl/PMFA}
 }
 
@@ -122,7 +122,7 @@ address = "Aachen",
 
 @InProceedings{pandas1,
   author    = { {W}es {M}c{K}inney },
-  title     = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
+  title     = { Data Structures for Statistical Computing in {P}ython },
   booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
   pages     = { 56 - 61 },
   year      = { 2010 },
@@ -143,7 +143,6 @@ year = {2025}
 @online{plotly,
 author = {{Plotly Technologies Inc.}},
 title = {Collaborative data science},
-publisher = {Plotly Technologies Inc.},
 address = {Montreal, QC},
 year = {2015},
 url = {https://plot.ly}
@@ -164,11 +163,12 @@ url = {https://plot.ly}
   year      = 2007
 }
 
-@online{remind-mfa,
-author = {Dürrwächter, Jakob and Hosak, Merlin and Weiß, Bennet and Zhang, Qianzhi and Ueckerdt, Falko},
+@software{remind-mfa,
+author = {Dürrwächter, Jakob and Hosak, Merlin and Weiß, Bennet and Schweiger, Leonie and Zhang, Qianzhi and Ueckerdt, Falko},
 title = {REMIND-MFA},
-year = {2025},
-url = {https://github.com/pik-piam/remind-mfa},
+year = {2026},
+version = {v0.2.0},
+url = {https://remind-mfa.readthedocs.io},
 }
 
 
@@ -180,7 +180,7 @@ url = {https://transience-eu-mfa.readthedocs.io}
 }
 
 @online{AutumnSchool2025Website,
-author = {Mutel, Chris},
+author = {{Départ de Sentier}},
 title = {DdS Autumn School 2025},
 year = {2025},
 month = jul,
@@ -188,57 +188,9 @@ url = {https://www.d-d-s.ch/schools/nov-25},
 urldate = {2026-01-19}
 }
 
-@online{ISIE_SEM_2024,
-title = {ISIE-SEM Summer School},
-year = {2024},
-month = jan,
-url = {https://is4ie.org/events/event/isie-sem-summer-school},
-urldate = {2026-01-19}
-}
-
-@online{AutumnSchool2025Github,
-title = {Schools-2025-November-Switzerland},
-year = {2026},
-month = jan,
-url = {https://github.com/Depart-de-Sentier/Schools-2025-November-Switzerland},
-urldate = {2026-01-19}
-}
-
-@online{PAW_MFA_LCA_2025,
-title = {PAW_MFA_LCA_2025},
-year = {2026},
-month = jan,
-url = {https://github.com/isabelapi/PAW_MFA_LCA_2025},
-urldate = {2026-01-19}
-}
-
-@online{DdS_REFLOC,
-title = {DdS_REFLOC},
-year = {2026},
-month = jan,
[…]
