+++ paper/paper.bib
@@ -1,7 +1,7 @@
 
 @article{odym,
     author = {Pauliuk, Stefan and Heeren, Niko},
-    title = {ODYM—An open software framework for studying dynamic material systems: Principles, implementation, and data structures},
+    title = {ODYM — {A}n open software framework for studying dynamic material systems: Principles, implementation, and data structures},
     journal = {Journal of Industrial Ecology},
     volume = {24},
     number = {3},
@@ -50,7 +50,7 @@ year = {2015}
 
 @software{pymfa2.1,
 author = {Thiébaud, Esther and Alexandru, Carol and Badat, Rolf and Kohler, Dimitri and Hilty, Lorenz},
-title = {pymfa2: Ein Werkzeug zur Analyse von Materialflüssen in Python 2.7},
+title = {{pymfa2: Ein Werkzeug zur Analyse von Materialflüssen in Python 2.7}},
 version = {2.1},
 year = {2019},
 url = {https://bitbucket.org/Xeelk/pymfa2/src/master/}
@@ -181,7 +181,7 @@ url = {https://transience-eu-mfa.readthedocs.io}
 
 @online{AutumnSchool2025Website,
 author = {{Départ de Sentier}},
-title = {DdS Autumn School 2025},
+title = {{DdS Autumn School 2025}},
 year = {2025},
 month = jul,
 url = {https://www.d-d-s.ch/schools/nov-25},
