+++ .pre-commit-config.yaml
@@ -1,13 +1,13 @@
 repos:
   - repo: https://github.com/mwouts/jupytext
-    rev: v1.19.0.dev0
+    rev: v1.19.1
     hooks:
     - id: jupytext
       args: [--sync, --pipe, black]
       additional_dependencies:
         - black==24.10.0 # should match hook
   - repo: https://github.com/psf/black-pre-commit-mirror
-    rev: 25.12.0
+    rev: 26.3.0
     hooks:
       - id: black
         language_version: python3.12
