+++ paper/paper.bib
@@ -13,26 +13,6 @@
     year = {2020}
 }
 
-@article{Lauinger21,
-	title = {A general framework for stock dynamics of populations and built and natural environments},
-	volume = {25},
-	copyright = {© 2021 The Authors. Journal of Industrial Ecology published by Wiley Periodicals LLC on behalf of Yale University},
-	issn = {1530-9290},
-	url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/jiec.13117},
-	doi = {10.1111/jiec.13117},
-	abstract = {Sustainable development involves a responsible management of the interactions between humans and their built and natural environment. From a physical perspective, the interactions can be characterized as stocks and flows of energy and matter within and between these spheres. Understanding the dynamics of the stocks is essential to enable their responsible management. A large number of independent disciplines study the dynamics of individual stocks with specific methods. The resulting fragmentation of methods hampers interdisciplinary learning, including the integration of more specialized discipline-specific models into more encompassing ones. Here, we develop a general mathematical framework for dynamic stock models based on balance, intrinsic, and model-approach equations. We use the framework to classify a variety of stock models from different disciplines and discuss their applicability. The framework provides a common language for the interdisciplinary analysis of coupled human–environment systems. This article met the requirements for a gold-gold JIE data openness badge described at http://jie.click/badges.},
-	language = {en},
-	number = {5},
-	urldate = {2025-06-11},
-	journal = {Journal of Industrial Ecology},
-	author = {Lauinger, Dirk and Billy, Romain G. and Vásquez, Felipe and Müller, Daniel B.},
-	year = {2021},
-	note = {\_eprint: https://onlinelibrary.wiley.com/doi/pdf/10.1111/jiec.13117},
-	keywords = {industrial ecology, population dynamics, demographic metabolism, material and energy flow analysis, socio-economic metabolism, stock dynamics},
-	pages = {1136--1146},
-	file = {Full Text PDF:C\:\\Users\\jakobdu\\Zotero\\storage\\LASAGGI8\\Lauinger et al. - 2021 - A general framework for stock dynamics of populations and built and natural environments.pdf:application/pdf;Snapshot:C\:\\Users\\jakobdu\\Zotero\\storage\\HQ2TQTFH\\jiec.html:text/html},
-}
-
 @article{Pauliuk2015,
 author = {Pauliuk, Stefan and Majeau-Bettez, Guillaume and Müller, Daniel B.},
 title = {A General System Structure and Accounting Framework for Socioeconomic Metabolism},
@@ -250,4 +230,13 @@ urldate = {2026-01-19}
   month   = jan,
   url     = {https://indico.d-d-s.ch/event/1/page/13-courses-for-beginners-and-advanced-users-sold-out},
   urldate = {2026-01-19}
+}
+
+@book{brunner_rechberger_2016,
+  title = {Handbook of Material Flow Analysis},
+  author = {Brunner, Paul H. and Rechberger, Helmut},
+  year = {2016},
+  edition = {2},
+  publisher = {CRC Press},
+  doi = {10.1201/9781315313450}
 }
