+++ paper/paper.md
@@ -28,22 +28,22 @@ bibliography: paper.bib
 
 # Summary
 
-Material flow analysis (MFA) is a core method in industrial ecology that tracks time-dependent material flows within a system, such as a national economy, across all life cycle stages. It also accounts for material accumulation in stocks, including materials embodied in products, assets, and infrastructure at a given time. MFA supports resource management, environmental impact assessment, and the evaluation of circular economy strategies, informing policy advice, urban and regional planning, and sustainable product design.
+Material flow analysis (MFA) is a core method in industrial ecology that tracks material flows within a system, such as a national economy, across different life cycle stages [@brunner_rechberger_2016]. It also accounts for material accumulation in stocks, including materials embodied in products, assets, and infrastructure at a given time. MFA supports resource management, environmental impact assessment, and the evaluation of circular economy strategies, informing policy-making, urban and regional planning, and sustainable product design.
 
-flodym (Flexible Open Dynamic Material Systems Model) is a library of objects and functions needed to build dynamic MFA. Mathematically, a large part of MFA translates to operations between multi-dimensional arrays.
-flodym implements the `FlodymArray` class, which internally manages operations of one or several such arrays. Objects representing flows, stocks, and parameters all inherit from this class. Stocks include lifetime models for dynamic stock modelling, i.e. for calculating the relation of material flows entering a stock and the mass and age structure of that stock over time. The whole MFA system is realized with an abstract parent class, which users can implement a subclass of. The parent class includes checking routines for mass balance consistency and non-negative flows. flodym includes functionality for efficient read-in and export via pandas [@pandas1], [@pandas2], as well as visualization routines.
+flodym (Flexible Open Dynamic Material Systems Model) is a Python library of objects and functions needed to build dynamic MFA models. From a mathematical perspective, MFA largely involves operations on multi-dimensional arrays.
+flodym implements the `FlodymArray` class, which internally manages operations of one or more such arrays. Objects representing flows, stocks, and parameters inherit from this class. Stocks include lifetime models for dynamic stock modelling, i.e. for calculating the relationship between material inflows to a stock and the resulting mass and age structure of that stock over time. The whole MFA system is realized with an abstract parent class, which users can subclass. The parent class includes routines for checking mass balance consistency and non-negative flows. flodym includes functionality for efficient data import and export via pandas [@pandas1; @pandas2], as well as visualization routines.
 
-flodym is based on the concepts of the Open Dynamic Material Systems Model (ODYM) [@odym]. It can be seen as a re-implementation with vastly expanded functionality and improved structuring. As a result, flodym enables users to write customized, flexible MFAs, built for maintainability and future expansion.
+flodym is based on the concepts of the Open Dynamic Material Systems Model (ODYM) [@odym]. It can be seen as a re-implementation with expanded functionality and improved structuring. As a result, flodym enables users to write customized, flexible MFAs, designed for maintainability and future extension.
 
 # Statement of need
 
-MFA provides quantitative insights into how materials are extracted, transformed, used, and retained within socio-economic systems over time. In academia, it enables research on resource efficiency, environmental impacts, and sustainability transitions. In industry, it informs decisions on material sourcing, product lifetimes, and circular economy strategies. For policy-makers and planners, MFA provides an evidence base for resource governance, waste and demand forecasting, and the design of interventions.
[…]
