### 2026-03-11 58aba30
minor paper language edits
