+++ paper/paper.md
@@ -31,9 +31,9 @@ bibliography: paper.bib
 Material flow analysis (MFA) is a core method in industrial ecology that tracks material flows within a system, such as a national economy, across different life cycle stages [@brunner_rechberger_2016]. It also accounts for material accumulation in stocks, including materials embodied in products, assets, and infrastructure at a given time. MFA supports resource management, environmental impact assessment, and the evaluation of circular economy strategies, informing policy-making, urban and regional planning, and sustainable product design.
 
 flodym (Flexible Open Dynamic Material Systems Model) is a Python library of objects and functions needed to build dynamic MFA models. From a mathematical perspective, MFA largely involves operations on multi-dimensional arrays.
-flodym implements the `FlodymArray` class, which internally manages operations of one or more such arrays. Objects representing flows, stocks, and parameters inherit from this class. Stocks include lifetime models for dynamic stock modelling, i.e. for calculating the relationship between material inflows to a stock and the resulting mass and age structure of that stock over time. The whole MFA system is realized with an abstract parent class, which users can subclass. The parent class includes routines for checking mass balance consistency and non-negative flows. flodym includes functionality for efficient data import and export via pandas [@pandas1; @pandas2], as well as visualization routines.
+flodym implements the `FlodymArray` class, which internally manages operations of one or more such arrays. Objects representing flows, stocks, and parameters inherit from this class. Stocks include lifetime models for dynamic stock modelling, i.e. for calculating the relationship between material inflows to a stock and the resulting mass and age structure of that stock over time. The whole MFA system is realized with an abstract parent class, which users can subclass. The parent class includes routines for checking mass balance consistency and non-negative flows. flodym offers functionality for efficient data import and export via pandas [@pandas1; @pandas2], as well as visualization routines.
 
-flodym is based on the concepts of the Open Dynamic Material Systems Model (ODYM) [@odym]. It can be seen as a re-implementation with expanded functionality and improved structuring. As a result, flodym enables users to write customized, flexible MFAs, designed for maintainability and future extension.
+flodym is based on the concepts of the Open Dynamic Material Systems Model (ODYM) [@odym]. It is a re-implementation with expanded functionality and improved structuring. As a result, flodym enables users to write customized, flexible MFAs, designed for maintainability and future extension.
 
 # Statement of need
 
@@ -42,7 +42,7 @@ MFA provides quantitative insights into how materials are extracted, transformed
 MFA is therefore widely used in industrial ecology and related fields. Several review papers document the development and applications of MFA, including @Muller2014Feb, @Bringezu2018Jan, @Graedel2019Nov, and @Streeck2023Apr.
 
 The broad use of MFA makes widely applicable and easily accessible MFA tools essential for a large audience in academia, industry, and policy-making.
-Compared to MFA software using a graphical user interface, a code library such as flodym allows customization to users' individual needs, which is vital in academia. This approach substantially expands the range of possible applications.
+Compared to MFA software using a graphical user interface, a code library such as flodym allows customization to users' individual needs, and substantially expands the range of possible applications, which is vital in academia.
 One example is the coupling with other modeling tools like life cycle assessment or integrated assessment models, where custom implementations enable flexible data exchange through programmatic interfaces.
 
 # State of the field
@@ -51,15 +51,15 @@ While there are several existing open source MFA software packages, ODYM [@odym]
 
 ## ODYM
 
-ODYM is widely used in the industrial ecology community and related research areas. One of ODYM's strengths is that it builds on an abstraction of the principles and structures of MFA:
+ODYM is widely used in the industrial ecology community and related research areas. One of its strengths is that it builds on an abstraction of the principles and structures of MFA:
 
[…]
