### 2026-03-11 3a343d6
adapt pre-commit hook to execute notebooks after conversion
