+++ .pre-commit-config.yaml
@@ -3,7 +3,7 @@ repos:
     rev: v1.19.0.dev0
     hooks:
     - id: jupytext
-      args: [--sync, --pipe, black]
+      args: [--sync, --execute, --pipe, black]
       additional_dependencies:
         - black==24.10.0 # should match hook
   - repo: https://github.com/psf/black-pre-commit-mirror
