### 2026-03-11 d670525
add jupyter client dependency to jupytext hook
