+++ .pre-commit-config.yaml
@@ -5,7 +5,8 @@ repos:
     - id: jupytext
       args: [--sync, --execute, --pipe, black]
       additional_dependencies:
-        - black==24.10.0 # should match hook
+        - black==25.12.0 # should match hook
+        - jupyter-client==8.8.0
   - repo: https://github.com/psf/black-pre-commit-mirror
     rev: 25.12.0
     hooks:
