### 2026-03-11 6251d4a
adapt jupyter dependency in jupytext hook
