### 2026-03-11 ac7d297
adapt jupyter dependency version
