+++ .pre-commit-config.yaml
@@ -6,7 +6,7 @@ repos:
       args: [--sync, --execute, --pipe, black]
       additional_dependencies:
         - black==25.12.0 # should match hook
-        - jupyter==1.1.1
+        - jupyter>=1.1.1
   - repo: https://github.com/psf/black-pre-commit-mirror
     rev: 25.12.0
     hooks:
