+++ .pre-commit-config.yaml
@@ -1,21 +1,13 @@
+ci:
+    autofix_commit_msg: "[pre-commit.ci] auto fixes"
+    autoupdate_schedule: monthly
+    skip: [] # List hooks to skip in CI if needed
+
 repos:
-  - repo: https://github.com/mwouts/jupytext
-    rev: v1.19.0.dev0
+  - repo: local
     hooks:
-    - id: jupytext
-      args: [--sync, --execute, --pipe, black]
-      additional_dependencies:
-        - black==25.12.0 # should match hook
-        - jupyter>=1.1.1
-  - repo: https://github.com/psf/black-pre-commit-mirror
-    rev: 25.12.0
-    hooks:
-      - id: black
-        language_version: python3.12
-  - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v6.0.0
-    hooks:
-      - id: check-yaml
-      - id: check-case-conflict
-      - id: end-of-file-fixer
-      - id: trailing-whitespace
+      - id: block-ipynb-in-protected-dirs
+        name: Protect examples and howtos
+        files: ^(examples|howtos)/.*\.ipynb$
+        entry: python -c "import sys; print('\n❌ ERROR: Direct edits to .ipynb in examples/ or howtos/ are forbidden.\n'); sys.exit(1)"
+        language: system
