+++ .github/workflows/main_actions.yml
@@ -1,6 +1,3 @@
-# This workflow will install Python dependencies, run tests and lint with a single version of Python
-# For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
-
 name: flodym.tests
 
 on:
@@ -10,21 +7,66 @@ on:
     branches: [ "main" ]
 
 permissions:
-  contents: read
+  contents: write # Needed to push synced notebooks back to the repo
 
 jobs:
-  build:
-
-    runs-on: windows-latest
+  build-and-test:
+    runs-on: ubuntu-latest
 
     steps:
-    - uses: actions/checkout@v4
+    - name: Checkout repository
+      uses: actions/checkout@v4
+      with:
+        fetch-depth: 0 # Ensures full history for the auto-commit step
+
     - name: Set up Python 3.12
-      uses: actions/setup-python@v3
+      uses: actions/setup-python@v5
       with:
         python-version: "3.12"
-    - name: Install dependencies and run tests
+        cache: 'pip' # Automatically caches your pip dependencies
+
+    - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
+        pip install black jupytext nbconvert ipykernel
+        # Install project + test dependencies
         pip install -e ".[tests]"
+
+    - name: Sync and Execute Notebooks
+      # This block handles the examples/ and howtos/ directories specifically
+      run: |
+        set -e
+        FOLDERS=("examples" "howtos")
+
+        for FOLDER in "${FOLDERS[@]}"; do
+          if [ -d "$FOLDER" ]; then
+            echo "::group::Processing folder: $FOLDER"
+            for py_file in "$FOLDER"/*.py; do
+              # Skip if no .py files exist in the folder
+              [ -e "$py_file" ] || continue
+
+              ipynb_file="${py_file%.py}.ipynb"
+              echo "Processing: $py_file -> $ipynb_file"
+
+              # 1. Format the source code
+              black "$py_file"
+
+              # 2. Force Sync .py to .ipynb (overwrites any manual edits)
+              jupytext --to ipynb "$py_file" -o "$ipynb_file"
+
+              # 3. Execute Notebook (Job fails if code crashes)
+              jupyter nbconvert --to notebook --execute --inplace "$ipynb_file"
+            done
+            echo "::endgroup::"
+          fi
+        done
+
+    - name: Commit and Push changes
+      uses: stefanzweifel/git-auto-commit-action@v5
+      with:
+        commit_message: "chore: auto-sync and execute notebooks [skip ci]"
+        file_pattern: "examples/*.ipynb howtos/*.ipynb examples/*.py howtos/*.py"
+
+    - name: Run Pytest
+      run: |
         python -m pytest
