+++ .pre-commit-config.yaml
@@ -4,8 +4,10 @@ repos:
       - id: block-ipynb-in-protected-dirs
         name: Protect examples and howtos
         files: ^(examples|howtos)/.*\.ipynb$
-        # The '|' allows us to write multi-line strings without escape errors
         entry: |
-          python -c "import sys; print('\n' + '='*50 + '\n❌ ERROR: DIRECT IPYNB COMMIT BLOCKED\n' + '='*50 + '\nLocation: ' + sys.argv[1] + '\n\nIn this repo, .py files are the source of truth.\n1. Discard changes to the .ipynb file.\n2. Edit the paired .py file instead.\n3. Push the .py; GitHub CI will sync/run the notebook.\n' + '='*50 + '\n'); sys.exit(1)"
+          python -c "import sys;
+          if len(sys.argv) > 1:
+              print('\n❌ ERROR: MANUAL IPYNB MODIFICATION DETECTED\n');
+              sys.exit(1)"
         language: system
         pass_filenames: true
