+++ examples/example2.py
@@ -59,7 +59,7 @@ from flodym import (
 from flodym.export import PlotlyArrayPlotter
 
 # needed only for correct rendering on the readthedocs homepage
-pio.renderers.default = "browser"
+# pio.renderers.default = "browser"
 
 # %% [markdown]
 # ## 2. Define the data requirements, flows, stocks and MFA system equations
