### 2026-03-11 911ac49
undo changes of relative example data paths
