### 2026-03-12 8b7dfb9
add run_notebooks script. add a test to verify they all return with a zero exit status. add a Makefile with commands to syn and/or run all notebooks. adapt the pre-commit config to do the syncing only in py->ipynb direction, without changing iypnb outputs, and only use black on the py files.
