+++ .pre-commit-config.yaml
@@ -1,16 +1,16 @@
 repos:
-  - repo: https://github.com/mwouts/jupytext
-    rev: v1.19.0.dev0
-    hooks:
-    - id: jupytext
-      args: [--sync, --pipe, black]
-      additional_dependencies:
-        - black==24.10.0 # should match hook
   - repo: https://github.com/psf/black-pre-commit-mirror
     rev: 25.12.0
     hooks:
       - id: black
         language_version: python3.12
+        exclude: \.ipynb$
+  - repo: https://github.com/mwouts/jupytext
+    rev: v1.19.0.dev0
+    hooks:
+      - id: jupytext
+        files: ^(examples|howtos)/.*\.py$
+        args: [--to, ipynb, --update]
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v6.0.0
     hooks:
