### 2026-03-12 63c23e6
make test_notebooks run on py files instead
