### 2026-03-12 67c4387
update dependencies for tests
