+++ pyproject.toml
@@ -32,6 +32,7 @@ examples = [
 tests = [
     "pytest>=8.3.2",
     "polyfactory>=2.16.2",
+    "ipython>=8.0.0",
 ]
 docs = [
     "sphinx>=7.4.0",
