### 2026-03-12 1b72740
update encoding in run_notebooks
