### 2026-03-12 0aa2f3e
add nbformat to tests deps
