+++ pyproject.toml
@@ -33,6 +33,7 @@ tests = [
     "pytest>=8.3.2",
     "polyfactory>=2.16.2",
     "ipython>=8.0.0",
+    "nbformat>=5.10.4",
 ]
 docs = [
     "sphinx>=7.4.0",
