### 2026-03-12 1f2f5cd
emphasize interactive functionality in plotly plot in example 2
