### 2026-03-12 fa9a7a6
change plotly plotting to be displayed on readthedocs
