### 2026-03-12 dc4b264
move note on interactive figures into subtitle
