+++ examples/example2.py
@@ -22,7 +22,7 @@
 #
 # * How much copper accumulates in the secondary steel assuming that all available scrap is remelted?
 # * How much manganese is lost in the remelting process assuming that all available scrap is remelted?
-# * What is more effective in reducing the copper concentraction of secondary steel: A reduction of the shredding yield factor for copper from EoL machines into steel scrap of 25% or an increase in the EoL buildings flow by 25%? (All other variables and parameters remaining equal)
+# * What is more effective in reducing the copper concentration of secondary steel: A reduction of the shredding yield factor for copper from EoL machines into steel scrap of 25% or an increase in the EoL buildings flow by 25%? (All other variables and parameters remaining equal)
 #
 # <img src="pictures/SteelAlloyElementsWasteMgt.png" width="554" height="490" alt="Simple MFA system">
 #
@@ -213,11 +213,13 @@ mfa_example.compute()
 # Here we answer the research questions from the beginning of the notebook.
 #
 # **How much copper accumulates in the secondary steel assuming that all available scrap is remelted?**
-#
-# <span style="color:#8B0000;"><strong>Clicking on the `Fe` entry of the plot legend hides it and adjusts the y-axis to better display the trace elements `Mn` and `Cu`.</strong></span>
 
 # %%
 remelted = mfa_example.flows["remelting => sysenv"]
+legend_hint_subtitle = (
+    "Clicking on the 'Fe' legend entry hides it and adjusts the y-axis to better display "
+    "the trace elements 'Mn' and 'Cu'."
+)
 
 plotter = PlotlyArrayPlotter(
     array=remelted,
@@ -226,6 +228,9 @@ plotter = PlotlyArrayPlotter(
     title="Remelted steel",
 )
 fig = plotter.plot()
+fig.update_layout(
+    title=dict(text=f"{fig.layout.title.text}<br><sup><i>{legend_hint_subtitle}</i></sup>")
+)
 fig.show(renderer="notebook")
 
 
@@ -239,6 +244,9 @@ plotter = PlotlyArrayPlotter(
     title="Share of copper and manganese in secondary steel",
 )
 fig = plotter.plot()
+fig.update_layout(
+    title=dict(text=f"{fig.layout.title.text}<br><sup><i>{legend_hint_subtitle}</i></sup>")
+)
 fig.show(renderer="notebook")
 
 
@@ -246,8 +254,6 @@ fig.show(renderer="notebook")
 # The copper flow in the secondary steel increases linearly from 0.34 kt/yr in 1980 to 0.78 kt/yr in 2010. The concentration of copper declines in a hyperbolic curve from 0.294% in 1980 to 0.233% in 2010.
 #
 # That concentration is below 0.4% at all times, the latter being the threshold for construction grade steel, but above 0.04%, which is the threshold for automotive steel.
-#
-# <span style="color:#8B0000;"><strong>Clicking on the `Fe` entry of the plot legend hides it and adjusts the y-axis to better display the trace elements `Mn` and `Cu`.</strong></span>
 
 # %% [markdown]
 # **How much manganese is lost in the remelting process assuming that all available scrap is remelted?**
