### 2026-03-12 382c2dd
hide Fe lines in plots by default
