### 2026-03-12 7361e87
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
