### 2026-03-12 #10105 @JakobBD
Dear Dear @TimoDiepers and @michaelweinold,

Thanks to both of you for the timely review and the very helpful remarks!
I have updated both paper and code.

In particular:

# Review by @TimoDiepers

For the issues on the code, see the linked PRs.

For the paper:

>     * Many of the points from the „State of the field“ section describe the features / design of flodym. I’d suggest to move these points (examples and descriptions of the simplifications flodym enables) to the software design section, and rather focus the state of the field section on the description (including potential downsides) of existing software. This would also resolve some of the redundancy that there currently is between the two sections.

Points moved, as suggested.

>     * Unify how courses are cited. For example, compare „ISIE SEM summer school“ and „Schools-2025-november-switzerland“. For the latter, you might also consider updating the link to point to the actual event (https://www.d-d-s.ch/schools/nov-25/)

Fixed.

>     * Language-related:
>
>       * "MFA provides quantitative insights _**into**_ how materials are extracted"
>       * "...not reached publication yet, but **_its_** external user base is growing rapidly."
>       * Product Names: Github -> GitHub, Github actions -> GitHub Actions, VSCode -> VS Code
>       * "flodym" is sometimes written in italics, sometimes not.

All corrected, thanks.

# Review by @michaelweinold

A general note:
To my understanding, most of the sections (and section titles) are required by the journal.
However, I was still able to follow all your suggestions.

> ## Comments on the Publication:
> ### General Comments
>
> The article is on the longer side for JOSS manuscripts. It should be shortened possible through the removal of:
>
>     * The "Software design" section, which currently includes mostly subjective and unhelpful phrasing, such as "Providing a coherent, as clean as possible API (...)" and "What’s [sic!] more, flodym features different levels of integration (...).

I have removed this version of the section. Following Timo's suggestion, I have moved parts of the "State of the field" section to the "Software design" section.

>     * Large parts of the "Research impact statement", which starts with the unhelpful statement that "flodym is a new tool and therefore applications have not reached publication yet". The list of "teaching events" should also include only past events. The mention of the number of GitHub stars, which is both [easily manipulated](https://medium.com/@vanshagrawal561/i-bought-fake-github-stars-so-you-dont-have-to-88f12c8eccb6) and (hopefully) subject to changes in the future should also be removed.

I have removed the sentences you suggested.

>
> ### Specific Comments
>
>     * Ensure you consistently name/stylize software packages. You write _flodym_, but also eg. pandas ([consistent with the project](https://pandas.pydata.org/about/)) and numpy ([inconsistent with the project](https://numpy.org/about/)).

I have removed the italics on flodym, and capitalized NumPy.

> ### Comments by Section
> #### Summary
>
> In general, this section is well written - but in "(...) including sanity checks for the system", what are those sanity checks - are they for non-negative flows, etc.?

Rephrased, adding specifics on these checks: "The parent class includes routines for checking mass balance consistency and non-negative flows."

> #### Statement of need
>
> I recommend removing the sentence including the publication count "Scopus lists 4260 publications (...)", since this number is meaningless on its own - and no year for the search was given.

Sentence removed.

> The unnecessary brackets in citations of the second paragraph needs to be removed: Instead of "(...) for example, (Müller et al., 2014), (Bringezu & Moriguchi, 2018), (...)", it should be "(...) for example, Müller et al., 2014, Bringezu & Moriguchi, 2018, (...)".

Fixed.

> In "This makes general and easily accessible MFA tools (...)" what would be a "general MFA tool"?

Rephrased to "widely applicable".

> The abbreviation "GUI" is introduced but never defined.

Fixed.

> The "L" in "(...) tools like Life cycle (...)" should not be capitalized.

Fixed.

> "(...) where data can be exchanged flexibly through a coded solution." sounds a bit awkward.

Rephrased to: "(...) where custom models enable flexible data exchange through a coded interface."

> #### State of the field
>
> While you mention a few other tools at the end of the section, you really only compare your tool to ODYM here... I would therefore suggest you rename this section to something including "ODYM".

JOSS prescribes this section title, but I have structured the section into two subsections, one of which is called ODYM.

> Ensure you consistently use punctuation in enumerations - currently only some of your bullet points end with a full stop.

Fixed.

> In your code examples, for instance, `waste = np.einsum('trp,pw->trw', end_of_life_products, waste_share)` - what does this do? You cannot assume the reader knows what `trp`, etc. are without having used ODYM. You might removing un-interesting paragraphs on plotting functionality while expanding more on the (syntax)-benefits of flowdym.

I have expanded these paragraphs, while heavily shortening the one on plotting functionality.

> #### Acknowledgements
>
> The authors mix Third-Person Objective ("We acknowledge...") and First-Person Active voice ("Thank you...").

Fixed.

> #### References
>
> References are a mess and need to be unified:
>
>     * Some items give metadata description (`[Computer software]`), others leave the reader guessing from the URL (`dds2025manure. (2026, January). https://github.com/gergosuto/dds2025manure`).

Fixed.
Note: I have moved some GitHub repository references to footnotes, which seemed more fitting in this context.

>     * Some items incorrectly use the _in proceedings_ descriptor "In", even though the reference is clearly to the homepage of an event ("Brightcon 2025, hackathon & courses in Grenoble and online. (2026). In Indico. (...)"

Fixed.

>     * Plotly Technologies Inc. is apparently so important it is mentioned twice in the same reference.

Fixed.

>     * Capitalization of reference titles must be fixed in many places. For example: "MFA using r" (the "R" is presumably the programming language)

Fixed.
