### 2026-03-12 #10105 @michaelweinold
@JakobBD, please just have the AI agent of your choice have a go at the references. Needs further cleaning. For instance:

- `grenoble` not capitalized.
- Journal of Industrial Ecology once abbreviated, once not abbreviated.
- Book `Environmental informatics and industrial ecology` should at least have an ISBN, if DOI is not available.
- Italics used inconsistently, sometimes indicating the work itself, sometimes indicating the journal, etc.

I am happy with all other modifications to the manuscript. Thank you for diligently implementing them.
