### 2026-03-12 46b0e84
paper references formatting
