+++ paper/paper.bib
@@ -2,7 +2,7 @@
 @article{odym,
     author = {Pauliuk, Stefan and Heeren, Niko},
     title = {ODYM — {A}n open software framework for studying dynamic material systems: Principles, implementation, and data structures},
-    journal = {Journal of Industrial Ecology},
+    journal = {J. Ind. Ecol.},
     volume = {24},
     number = {3},
     pages = {446-458},
@@ -16,7 +16,7 @@
 @article{Pauliuk2015,
 author = {Pauliuk, Stefan and Majeau-Bettez, Guillaume and Müller, Daniel B.},
 title = {A General System Structure and Accounting Framework for Socioeconomic Metabolism},
-journal = {Journal of Industrial Ecology},
+journal = {J. Ind. Ecol.},
 volume = {19},
 number = {5},
 pages = {728-741},
@@ -40,13 +40,14 @@ url = {https://bitbucket.org/Xeelk/pymfa2/src/master/}
 @software{PMFA,
 author = {Kawecki-Wenger, Delphine},
 title = {PMFA: Base functions for performing a Probabilistic MFA using {R}},
+year = {2019},
 url = {https://github.com/empa-tsl/PMFA}
 }
 
 @article{OMAT,
 author = {Villalba, Gara and Hoekman, Paul},
 title = {Using Web-Based Technology to Bring Hands-On Urban Material Flow Analysis to the Classroom},
-journal = {Journal of Industrial Ecology},
+journal = {J. Ind. Ecol.},
 volume = {22},
 number = {2},
 pages = {434-442},
@@ -64,9 +65,10 @@ year = 2008,
 booktitle = "Environmental Informatics and Industrial Ecology",
 publisher = "Shaker Verlag",
 address = "Aachen",
+isbn = "978-3832273132"
 }
 
-@Article{numpy,
+@article{numpy,
  title         = {Array programming with {NumPy}},
  author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
                  van der Walt and Ralf Gommers and Pauli Virtanen and David
@@ -100,7 +102,7 @@ address = "Aachen",
     url          = {https://doi.org/10.5281/zenodo.3509134}
 }
 
-@InProceedings{pandas1,
+@inproceedings{pandas1,
   author    = { {W}es {M}c{K}inney },
   title     = { Data Structures for Statistical Computing in {P}ython },
   booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
@@ -128,7 +130,7 @@ year = {2015},
 url = {https://plot.ly}
 }
 
-@Article{pyplot,
+@article{pyplot,
   Author    = {Hunter, J. D.},
   Title     = {Matplotlib: A 2D graphics environment},
   Journal   = {Computing in Science \& Engineering},
@@ -186,7 +188,6 @@ urldate = {2026-01-19}
 	author = {Bringezu, Stefan and Moriguchi, Yuichi},
 	title = {{Material flow analysis}},
 	booktitle = {{Green Accounting}},
-	journal = {Taylor {\&} Francis},
 	pages = {149--166},
 	year = {2018},
 	month = jan,
@@ -194,6 +195,7 @@ urldate = {2026-01-19}
 	address = {New York, NY, USA},
 	doi = {10.4324/9781315197715-6}
 }
+
 
 @article{Graedel2019Nov,
 	author = {Graedel, Thomas E.},
@@ -224,7 +226,7 @@ urldate = {2026-01-19}
 }
 
 @online{Brightcon,
-  title   = {Brightcon 2025: Hackathon \& Courses in Grenoble and Online},
+  title   = {Brightcon 2025: Hackathon \& Courses in {G}renoble and Online},
   author  = {{Départ de Sentier}},
   year    = {2026},
   month   = jan,
