### 2026-03-12 28660d9
more paper references formatting
