+++ paper/paper.bib
@@ -5,7 +5,7 @@
     journal = {J. Ind. Ecol.},
     volume = {24},
     number = {3},
-    pages = {446-458},
+  pages = {446--458},
     keywords = {dynamic modeling, material flow analysis (MFA), open science, Python, software, substance flow analysis (SFA)},
     doi = {10.1111/jiec.12952},
     url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/jiec.12952},
@@ -19,7 +19,7 @@ title = {A General System Structure and Accounting Framework for Socioeconomic M
 journal = {J. Ind. Ecol.},
 volume = {19},
 number = {5},
-pages = {728-741},
+pages = {728--741},
 keywords = {accounting, bipartite graphs, industrial ecology model, societal metabolism, supply and use tables, system of national accounts (SNA)},
 doi = {10.1111/jiec.12306},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/jiec.12306},
@@ -50,7 +50,7 @@ title = {Using Web-Based Technology to Bring Hands-On Urban Material Flow Analys
 journal = {J. Ind. Ecol.},
 volume = {22},
 number = {2},
-pages = {434-442},
+pages = {434--442},
 keywords = {education and training, software, learning by doing, economy-wide material flow analysis (EW-MFA), urban metabolism, environmental education},
 doi = {10.1111/jiec.12553},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/jiec.12553},
@@ -63,6 +63,7 @@ author = "Cencic, Oliver and Rechberger, Helmut",
 title = "Material Flow Analysis with Software STAN",
 year = 2008,
 booktitle = "Environmental Informatics and Industrial Ecology",
+editor = "Moeller, Andreas and Page, Bernd and Schreiber, Martin",
 publisher = "Shaker Verlag",
 address = "Aachen",
 isbn = "978-3832273132"
@@ -106,7 +107,7 @@ isbn = "978-3832273132"
   author    = { {W}es {M}c{K}inney },
   title     = { Data Structures for Statistical Computing in {P}ython },
   booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
-  pages     = { 56 - 61 },
+  pages     = {56--61},
   year      = { 2010 },
   editor    = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
   doi       = { 10.25080/Majora-92bf1922-00a }
@@ -115,14 +116,14 @@ isbn = "978-3832273132"
 
 @software{pydantic,
 title = {Pydantic},
-author = { Colvin , Samuel and Jolibois , Eric and Ramezani , Hasan and Garcia Badaracco , Adrian and Dorsey , Terrence and Montague , David and Matveenko , Serge and Trylesinski , Marcelo and Runkle , Sydney and Hewitt , David and Hall , Alex and Plot , Victorien},
-repository-code = {https://github.com/pydantic/pydantic},
+author = {Colvin, Samuel and Jolibois, Eric and Ramezani, Hasan and Garcia Badaracco, Adrian and Dorsey, Terrence and Montague, David and Matveenko, Serge and Trylesinski, Marcelo and Runkle, Sydney and Hewitt, David and Hall, Alex and Plot, Victorien},
+note = {Source code: https://github.com/pydantic/pydantic},
 url = {https://docs.pydantic.dev/latest/},
 version = {v2.11.7},
 year = {2025}
 }
 
-@online{plotly,
+@software{plotly,
 author = {{Plotly Technologies Inc.}},
 title = {Collaborative data science},
[…]
