+++ paper/paper.bib
@@ -5,7 +5,7 @@
     journal = {J. Ind. Ecol.},
     volume = {24},
     number = {3},
-  pages = {446--458},
+    pages = {446--458},
     keywords = {dynamic modeling, material flow analysis (MFA), open science, Python, software, substance flow analysis (SFA)},
     doi = {10.1111/jiec.12952},
     url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/jiec.12952},
