### 2026-03-12 #10105 @JakobBD
Thanks! I have fixed your first three points.
I have given an AI agent the bib-file, and a copy-paste of the compiled version.
It has found some further minor inconsistencies. For example, I've added editors to two books.

I'm afraid the italics inconsistency still persists.
I suppose it is caused by the pandoc/LaTeX bibliography style in the paper generation GitHub Action.
From what I see, the logic is: If there is a journal/book etc, make that italic, else take the work itself (e.g., for  GitHub repos).
There are further differences in formatting which I assume are caused by different item types (journal article vs online resource and so on).
I don't see how I could change that, it's controlled by JOSS.

Also, I have sometimes chosen to keep citations as given by the authors rather than harmonizing them (e.g., version numbers with or without a leading "v").

Let me know if you think this is sufficient.
