### 2026-03-12 #10105 @michaelweinold
@marlinarnz, I can now recommend the article for publication. I thank @JakobBD for his patience and his diligent work. The software is a welcome contribution to the sustainability assessment ecosystem and the associated manuscript has now been greatly improved in quality.
