### 2026-03-13 b18b50f
add inplace option to cast_to method. adapt array stacking method to build on it
