### 2026-03-16 #10105 @marlinarnz
@JakobBD could you address the following ToDos prior to acceptance?

- [x] In your article, p.3, l. 121: What are "Python’s magic methods"? I sometimes feel the same but that does not sound very scientific.
- [x] Release a new version of flodym, if necessary, and set the version connected to this publication with `@editorialbot set <version here> as version`.
- [x] Archive this version in an external repository (e.g. Zenodo) and set this archive's DOI with `@editorialbot set <DOI here> as archive`. Make sure this archived version's author list, title, ORCIDs and its license match the JOSS article.
