### 2026-03-16 034c2c8
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/psf/black-pre-commit-mirror: 25.12.0 → 26.3.1](https://github.com/psf/black-pre-commit-mirror/compare/25.12.0...26.3.1)
- [github.com/mwouts/jupytext: v1.19.0.dev0 → v1.19.1](https://github.com/mwouts/jupytext/compare/v1.19.0.dev0...v1.19.1)
