+++ .pre-commit-config.yaml
@@ -1,12 +1,12 @@
 repos:
   - repo: https://github.com/psf/black-pre-commit-mirror
-    rev: 25.12.0
+    rev: 26.3.1
     hooks:
       - id: black
         language_version: python3.12
         exclude: \.ipynb$
   - repo: https://github.com/mwouts/jupytext
-    rev: v1.19.0.dev0
+    rev: v1.19.1
     hooks:
       - id: jupytext
         files: ^(examples|howtos)/.*\.py$
