### 2026-03-16 8c69cfc
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
