### 2026-03-17 #10105 @JakobBD
Dear @marlinarnz

Thanks for your support! I did the following:
- rephrased "magic methods" to: "Python’s special methods (methods with names in double underscores allowing operator overloading)"
- double checked and adapted authors, affiliations and ORCIDs
- made a release
- put everything on zenodo. Made sure that title, authors and license match.
