### 2026-03-17 #10105 @JakobBD
@marlinarnz I suppose you have to update version and DOI. You can just copy them from my comment. Thanks!
