### 2026-03-18 b5cfb0e
bugfixes for array stacking method
