+++ flodym/flodym_array_helper.py
@@ -22,7 +22,7 @@ def flodym_array_stack(flodym_arrays: list[FlodymArray], dimension: Dimension, a
     if dimension.letter in flodym_array.dims:
         raise ValueError(f"Dimension {dimension.letter} already present in FlodymArrays to stack")
     flodym_array0 = flodym_arrays[0]
-    extended_dimensions = flodym_array0.dims.insert(index=axis, dimension=dimension)
+    extended_dimensions = flodym_array0.dims.insert(index=axis, new_dim=dimension)
     extended = flodym_array0.cast_to(extended_dimensions, inplace=inplace)
     for item, flodym_array in zip(dimension.items, flodym_arrays):
         extended[{dimension.letter: item}] = flodym_array
