### 2026-03-18 3f3dd84
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
